Get better at finding Berkeley DB.
* configure.in: Use new macro SVN_LIB_BERKELEY_DB.  If we can't find
it, then just don't build the filesystem library.  We should also make
this turn off mod_dav_svn, but I'll leave that to Greg.
* subversion/Makefile.am (SUBDIRS): Build in libsvn_fs only if the configure
script says we should.
* ac-helpers/berkeley-db.m4 (SVN_CHECK_BERKELEY_DB): Replaced by:
(SVN_LIB_BERKELEY_DB): New macro.  Check --with-berkeley-db flag.  If
the user doesn't explicitly tell us where to look, search for Berkeley
DB in a number of standard places.  Cache results in a way that works
accurately even when the user tells us to look in a different place, or
not to look for Berkeley DB at all.