Clean up `const' usage.
This is a matter of style, but I think `const' should be used when the
additional static checking it enables accurately reflects functions'
contracts and structures' invariants.  When using `const' would
introduce casts which don't reflect genuine, deliberate type trickery
(which, given the rest of the C language, is a fine thing in many
cases) on the part of the code doing the cast, it's best to just avoid
it.  (This should be in a comment somewhere, not a log message, but
I'm not sure exactly where.)

* libsvn_fs/dbt.c (svn_fs__skel_to_dbt): Remove `const' from SKEL
argument.  This declares only the top skel of the tree to be constant,
whereas what we want to say is that the entire tree will be unchanged.
* libsvn_fs/dbt.h (svn_fs__skel_to_dbt): Same for declaration.
* libsvn_fs/skel.c (estimate_unparsed_size, unparse, use_implicit,
svn_fs__unparse_skel): Same.
(parse, list, implicit_atom, explicit_atom, svn_fs_parse_skel,
svn_fs__str_atom, svn_fs__mem_atom): Remove const from DATA and STR
arguments, and locals that point into DATA; the `data' member of a
skel isn't const, so this is misleading.
* libsvn_fs/skel.h (struct skel_t): Make `data' member non-const;
you're supposed to be able to munge the data a skel points to.
(svn_fs__parse_skel, svn_fs__str_atom, svn_fs__mem_atom,
svn_fs__unparse_skel): Fix declarations.
* libsvn_fs/clones-table.c (svn_fs__record_rename): Copy entry_name,
to keep the const promise.
* libsvn_fs/tests/skel-test.c (parse_cstr): Remove `const' qualifier.
Instead, cast the call sites; only the callers know that they will not
change the text of the string.