Add some important parts to RA interface that we forgot, given the way that editors work.  Without these things, replace_root() would be meaningless.
(get_commit_editor): the client must specify a ROOT_PATH in repository
that it plans to commit to.  The EDIT_BATON handed back with the
editor has this knowledge.

(do_update): similarly, the client must supply a ROOT_PATH in the
repository that represents the topmost subdir in the working copy that
is being updated.  The REPORT_BATON handed back should have this
knowledge.