autoconf output is supposed to be private. Otherwise, it may conflict with other autoconf'd programs.
Rename config.h to keep this point in mind, and to prevent filename conflict
with the (generically-named) config.h. Also remove the thing from
include/svn_path.h -- including on it in there means that everybody must
have it available, which just isn't true (specifically: it isn't shipped as
part of an SDK).

* configure.in: rename the header to "svn_private_config.h"
* autogen.sh: change an echo to reflect the new name
* svn_config.dsp: change for the new name
* client/diff.c, libsvn_wc/get_editor.c, mod_dav_svn/mod_dav_svn.c: change
    include file for the new name
* include/svn_path.h: remove include of "config.h"
  shift SVN_PATH_*_SEPARATOR constants into libsvn_subr/path.c; these are
    private values for the path code. carry comments along.
* libsvn_subr/path.c: constants and comments moved from svn_path.h

* svn_private_config.hw: added as a copy of config.hw
* config.hw: no longer applicable