Clarify the lifetime issues around the svn_txdelta_window_t structure. In some cases, it was created as a "free standing" object in its own subpool and svn_txdelta_free_window() needed to be called. Other times, it was a structure on the stack.
Now, it is clearer that the code creating the window is responsible
for cleaning it, and how they clean it. If svn_txdelta_next_window()
is used, then it is passed a pool, so the caller knows to clear the
pool to toss the window. If the caller constructs a window manually,
then they also know how to clean it up.

NOTE: this change may also dramatically fix our memory consumption. I
found that the vdelta hash tables were allocated in the stream's pool,
rather than for each window. I'm not positive what the resulting peak
was, but it appeared to be on the order of a pointer for each byte in
the source *and* target files that were being diff'd. (e.g. compare
two one megabyte files and you'd get 8 megabytes of pointer data).
Basically, the wrong pool was passed to svn_txdelta__vdelta().

* include/svn_delta.h (svn_txdelta_window_t): Removed the 'pool'
    member field.
  (svn_txdelta_next_window): take a POOL argument. update doc.
  (svn_txdelta_free_window): removed

* libsvn_fs/reps_strings.c (svn_fs__rep_deltify): allocate a subpool
    to contain the windows fetched via svn_txdelta_next_window. clear
    on each loop.

* tests/libsvn_delta/vdelta-test.c (main): allocate a subpool to
    contains the windows fetched via svn_txdelta_next_window. clear on
    each loop.

* libsvn_delta/delta.h (svn_txdelta__insert_op): add a POOL argument.

* libsvn_delta/vdelta.c (vdelta): pass pool argument to the
    svn_txdelta__insert_op() function for allocating space for the new
    operation.

* libsvn_delta/svndiff.c (write_handler): allocate the window on the
    stack, rather than on the heap. make the destruction of the
    subpool explicit, rather than via svn_txdelta_free_window().

* libsvn_delta/text_delta.c (svn_txdelta__make_window): place the new
    window directly into the pool passed, rather than into a new
    subpool.
  (svn_txdelta__insert_op): take a new POOL argument for allocations
    related to the inserted operations.
  (svn_txdelta_next_window): take a POOL argument for where to
    allocate the new window (rather than via a subpool of the
    stream->pool passed to make_window). create errors via this pool,
    rather than the stream's pool. use this pool for computing vdeltas
    instead of the stream pool(!!)
  (svn_txdelta_free_window): remove
  (svn_txdelta_send_string): remove assignment of the window's pool.
  (svn_txdelta_send_txstream): use a subpool for storing the windows,
    and clear on each iteration.

* libsvn_ra_dav/fetch.c (fetch_file_reader): remove assignment of the
    window's pool.