To satisfy those who want their ChangeLog's in GNU style rather than svn style, here's a script to munge the output of 'svn log' into something that approaches GNU ChangeLog style.
If someone is really motivated, they could grab some code from cvs2cl.pl and 
make this do line wrapping and formatting, but this is at least a start.

This closes (barely) issue #610, although there are several useful suggestions
in the issue that a motivated individual could take and add to this script.

* tools/dev/gnuify-changelog.pl
  new file.