Abstract the process of getting a node id's predecessor.
I'm no longer sure that the other abstractions mentioned or implied in
revision 1864's log message are worth the effort -- in particular,
abstracting the ID branching code would be tricky.  But maybe it'll
seem like a good idea again later?  In any case it's better not to do
it all at once.

* subversion/libsvn_fs/id.h, subversion/libsvn_fs/id.c
  (svn_fs__precede_id): New function, tweaks the ID in place.
  (svn_fs__id_predecessor): Just use above.  Return an empty ID
  instead of NULL.  Redocument.

* subversion/libsvn_fs/tree.c
  (txn_body_revisions_changed): Use svn_fs__precede_id instead of
  inline code.  Undo revision 1864's change to the `while' condition,
  since svn_fs__precede_id now handles that case implicitly.

* subversion/libsvn_fs/id.c
  (svn_fs_check_related): Use svn_fs__precede_id instead of inline
  code.

* subversion/libsvn_fs/deltify.c
  (svn_fs__stable_node): Adjust for new return value of
  svn_fs__id_predecessor.