This is the first part of the changes for issue 614, eg. to make svn use ISO-8601 timestamps.
It makes svn_time_from_nts understand a new style of timestamps,
roughly looking like this:

  2002-05-13T19:00:50.966679Z

For now, the code still generates old style timestamps, but that is to
be fixed after the next tarball is out.

This change should be completely non-intrusive - you should see no
change in behaviour. The only change which possibly might bite is the
change of using apr_time_exp_gmt instead of apr_time_exp_lt - but that
should not matter if APR has no bugs.

* subversion/libsvn_subr/time.c (timestamp_format): Change format to
  the new format.
  (old_timestamp_format): New variable.
  (svn_time_to_nts): Use apr_time_exp_gmt to get the time in GMT.
  Add commented version for printing the new format, but use the old
  format still.
  (svn_time_from_nts): Make it understand both new and old formats.

* subversion/tests/libsvn_subr/time-test.c: New file, tests for the
  time conversion functions, both new and old format.

* build.conf (time-test): Added building of the new test.