The beginning of an example: implement the 'svnlook' command entirely in Python. While we wouldn't require users to have Python installed, this does provide a platform for experimentation, and also a basis for how to write a commit-email script entirely in Python which binds directly to the SVN libraries (rather than invoking svnlook).
This implements the following commands: default, author, date,
dirs-changed, info, and log. It does not (yet) implement: changed,
diff, ids, and tree.

Note that one of the editor-based functions (dirs-changed) has been
implemented. The non-editor functions were cake :-) Given that one
editor-based function worked, I figured it was time to checkpoint. :-)

(some of the code is modelled directly from svnlook/main.c, but it is
 also trying to stick to a bit of "pythonic" style)