Fix issue #683:  'bug in repos_to_repos_copy() logic'.
In particular, we were failing when trying to resurrect deleted items
like this: 'svn cp -rN src_url dst_url', where src_url == dst_url.  In
this case, get_longest_ancestor(src_url, dst_url) was returning the
*very* same url as the hypothetical "parent" of the two; and then we'd
open the ra_sesson to that same url, and poof, we couldn't do the
copy.  The trick is to look for this edge case and make sure we open
an ra_session directly *above* the identical urls.

* libsvn_client/copy.c (repos_to_repos_copy): fix possible segfault
  from dereferencing NULL.  Also, look for special edge-case where
  src_url == dst_url, and set common parent of the urls to the actual
  *parent* of them.

* copy_tests.py 
  (resurrect_deleted_dir): new regression test for issue #683... using
  the new python test API!