Continuing work on issue #494.
This incorporates more of Marcus Comstedt's UTF-8 patch, by adding his
conversion routines to Subversion.  No one calls the new routines yet;
that will be added in a separate commit.

Note that this applies Marcus's `--enable-utf8' flag to configure.in.
If you build Subversion without the flag, the conversion routines
become nearly no-ops (they just check for non-ascii characters).  I
suspect this flag should eventually go away, but it's very handy right
now, because it's effectively an in-code branch -- meaning we don't
need to make a real branch in the repository for these changes; if
something's broken, it won't affect people who aren't working UTF-8
support.  Avoiding a branch in the repository is highly desirable,
because Marcus's changes touch a lot of code and would easily go stale
on a branch.

* subversion/include/svn_utf.h, subversion/libsvn_subr/utf.c: 
  New files.  Marcus, I tweaked check_non_ascii() a bit, and also
  documented some functions that lacked docs.  Please review if you
  can, thanks.

* subversion/libsvn_subr/libsvn_subr.dsp: Add utf.c.

* configure.in: Add --enable-utf8 option.