* tools/dev/svn-dev.el   (svn-log-path-derive): Fix edge case buglet, thanks to Blair Zajac   for noticing.
Note that this code is a branch from code in my .emacs.  This copy is
maintained for other developers (Ben C-S and Mike P, anyone else?), so
if it's not doing the Right Thing for them, it should be tweaked until
it is.  Meanwhile, I've fixed this bug in my .emacs as well :-).