Move Berkeley DB configuration from the code into the DB_CONFIG file, and change the way we open the database environment in preparation for automatic recovery.
* subversion/libsvn_fs/fs.c (allocate_env): Don't call set_lg_bsize.
(svn_fs_create_berkeley): Set it in the default DB_CONFIG file
instead.  Move the comment from allocate_env into the DB_CONFIG file,
and bump the default in-memory log buffer size to 256K, as it was done
in allocate_env.
Oh, and don't call strlen to get the length of a constant string.
(svn_fs_open_berkeley): Always pass the DB_CREATE flag to env->open,
so that the shared memory environment files are recreated as
necessary.

This means it is no longer necessary to call db_recover with the -e
flag when recovering Subversion databases, which brings us to:

* tools/backup/hot-backup.py: Call db_recover on the archived database
to put it in a consistent state and remove the shared-mem environment
files. This saves some space in the backup, and makes it ready for
immediate use.