There's some brokenness in our Python test suite's output-matching, discovered thanks to issue #838 and revision 2910.  This commit is just to clean up that brokenness, or some of it anyway, without fixing any of the newly-exposed bugs in Subversion itself.
I'm commenting out the now-failing tests, because developers depend on
'make check' to tell whether or not their changes have introduced new
bugs, and we don't want to interfere with that.  The way to deal with
the bugs revealed below is to file them in the issue tracker or fix
them, and reactive the regression tests when they pass again -- just
as we would never commit a live failing test without committing the
fix at the same time.

* subversion/tests/clients/cmdline/svntest/tree.py
  (build_tree_from_checkout): Match the exact number of spaces
  expected, not just any random amount of whitespace.  Note that this
  causes merge_tests.py:add_with_history to fail, so...

* subversion/tests/clients/cmdline/merge_tests.py 
  (test_list): Temporarily comment out add_with_history.

* subversion/tests/clients/cmdline/module_tests.py 
  (test_list): Temporarily comment out update_lose_external.  This
  test has been broken since at least revision 2849.

* build.conf
  (fs-test-scripts): Run module_tests.py too.  It wasn't run before
  because at the time I mistakenly thought tests listed here would be
  run over dav, and this test is known not to work over dav right
  now.  But we test over dav by hand anyway.