Follow up to revision 2973:
Simplify the error declaration interface, and force developers to
specify an offset when adding a new error.  This makes it harder to
accidentally insert a new error into the middle of a range:
presumably, while renumbering the subsequent errors, one would think
to oneself "There's got to be a better way, hmmm... perhaps I'll go
read the comments in this file to see if I'm missing something."
That's the idea, anyway.

* subversion/include/svn_error_codes.h 
  (SVN_ERR_CATEGORY_FIRST): Removed, all callers changed.
  (SVN_ERRDEF): Take an offset argument; all callers changed.