Working with Mike Pilato, fix part of issue #1000:
Ensure that there is no point in an update or switch when an entries
file is inaccurate, by making certain entry url changes happen in sync
with the edit, instead of relying on svn_wc__do_update_cleanup().
Note that this does not address the wcprop problems described in the
issue, only the entry problems.

* subversion/libsvn_wc/update_editor.c
  (struct dir_baton, struct file_baton): Rename `URL' to `new_URL'.
  (struct bump_dir_info): Add new_URL field.
  (maybe_bump_dir_info): Replaces maybe_bump_dir_revision, now handles
    new_URL as well, and tweaks the directory unconditionally.
  (get_entry_url): New helper function.
  (make_dir_baton): Fix new_URL calculation logic, and store the
    new_URL in the bump_dir_info structure too.  Use new helper
    get_entry_url.  Also, remove some truly redundant comments :-).
  (make_file_baton): Fix new_URL calculation logic.  Use new helper
    get_entry_url.
  (close_file): Pass fb->new_URL to svn_wc_install_file.