With Mike Pilato, follow up to revision 5459: Allow null date and/or author in the commit info returned from the server.
The problem with rev 5459 was when a file was committed, the server
would send back "" instead of true null for the author.  This would
get recorded as the author, so afterwards 'svn st -uv' would print the
empty string instead of "?".  But since the empty string doesn't look
like a column, the test suite would misparse the status output -- it
would take the filename as the author, and then claim that there was
no filename.  Badness.

* subversion/svnserve/serve.c
  (commit): Use optional tuple for date and author, instead of
    masking with the empty string.

* subversion/libsvn_ra_svn/client.c
  (ra_svn_end_commit): Expect optional tuple for date and author.