Followup to r5516, issue #1090 ("'svn cp URL wc' must notice foreign UUID").
In r5516, in the directory case, we decided that a foreign UUID meant
"don't do a checkout and add-with-history, but instead do an export
and straightforward add."  But this logic is flawed;  'svn export'
doesn't get properties.

So for now, in the foreign-UUID directory case, we're going to bail
and return SVN_ERR_UNSUPPORTED_FEATURE.  See the comment in the code.

* libsvn_client/copy.c (repos_to_wc_copy): don't support a dirURL->wc
  copy if the UUID is foreign.

* copy_tests.py (repos_to_wc):  make regression test expect failure in
  the directory case.