Fix whitespace-trimming functions.
Fix svn_stringbuf_strip_whitespace: it did not trim a string consisting only
of whitespace.  Remove svn_string_strip_whitespace: it had that bug and also
failed to set a null terminator, violating its invariants, and was not used,
and was not an appropriate operation to perform on a constant string.

* subversion/include/svn_string.h
* subversion/libsvn_subr/svn_string.c
  (svn_string_strip_whitespace): Removed.
  (svn_stringbuf_strip_whitespace): Strip even if string is all whitespace.