Do one of two compatibility cleanups from issue #1275:
In revision 6749, we taught ra-dav clients to DELETE activities they
no longer needed.  However, servers only learned how to support this
at the same time.  So, a client trying to DELETE an activity from an
older server would get a 500 HTTP_INTERNAL_SERVER_ERROR, and would
ignore that error.

After the new server code has been "out there" for a bit, though, we
want to yank the special code from delete_activity() which allows
status=500 to pass through as a non-errorful condition.  500s are,
generally speaking, most definitely the kinds of errors we want to
report.

* subversion/libsvn_ra_dav/commit.c
  (delete_activity): Don't special-case error 500,
    HTTP_INTERNAL_SERVER_ERROR, anymore.  Instead, ignore the result
    code of the request and just return any error.

* subversion/libsvn_ra_dav/ra_dav.h, subversion/libsvn_ra_dav/util.c
  (svn_ra_dav__request_dispatch): Change parameter's name to 'code_p',
     and allow it to be null, to better support the above changes.

Thanks to Mike Pilato for eyes.