With Mike Pilato, do the other compatibility cleanup from issue #1275:
The solution to issue #891 ("REPORT fails if item missing in HEAD")
was to create a new style of <update-report> that is always run
against the VCC, with both paths/revs meant for svn_repos_dir_delta
included in the request body.

This change makes the client stop falling back to old-style reports
(for old servers), and the server stop supporting them (for old
clients).

* subversion/libsvn_ra_dav/fetch.c
  (reporter_finish_report): If running an update report on the VCC
    failed, then just error, instead of falling back to old-style
    report.  Also, rearrange the code to get the VCC before doing
    anything else, just for clarity.

* subversion/mod_dav_svn/update.c
  (dav_svn__update_report): Only accept REPORTs against VCCs now.