Enable "diff -rCOMMITTED".  Make error reporting simpler and more consistent.
Although the revision specified as "COMMITTED" is the same as that specified
as "BASE", if it is specified as "COMMITTED" it will be retrieved from the
repository whereas if it is specified as "BASE" it will be retrieved from the
WC administrative area.  This is an implementation detail which should go away
one day.

* subversion/libsvn_client/diff.c (do_diff)
  Don't reject "COMMITTED".  It works and is consistent with other commands.
  When a URL is given, don't reject "BASE" or WC revision specifiers specially,
    because they will get rejected further down the line with a similar message
    anyway, in the same way that "COMMITTED" and "PREV" will.