Fix two bugs when the prompt providers are not installed (e.g. when the cmdline option --non-interactive is used):
  1. No username was used by ra_local.

  2. Authentication failed if --password was used without --username.

Move the code to get the username from the OS from the username and
simple prompt providers to the regular providers.  Improve code
readability and error checking.

The old code saved the values that was read from disk in the provider
baton to avoid saving unchanged data, but now that we have the
may_save parameter in the creds struct we use that method in all
simple and username providers instead.

Patch by: Ben Reser <ben@reser.org>
(Tweaked by me.)

* subversion/libsvn_client/username_providers.c

  (provider_baton_t): Remove redundant member.

  (get_creds): Remove.

  (save_creds): Remove.

  (username_first_creds): Move the code from get_creds here.  Move code
  to get the username from the OS in prompt_for_username_creds here.

  (username_save_creds): Move the code from save_creds here.

  (prompt_for_username_creds): Move code to get the username
  from the OS to username_first_creds.

* subversion/libsvn_client/simple_providers.c

  (provider_baton_t): Remove redundant members.

  (get_os_username): New function to get the OS username.

  (get_creds): Remove.

  (save_creds): Remove.

  (simple_first_creds): Move the code from get_creds here.  Get the
  default username from the OS.

  (simple_save_creds): Move the code from save_creds here.

  (prompt_for_simple_creds): Use get_os_username.