A major step towards making subcommands follow copy-history. (issue 1093). [Patch started by cmpilato, continued by me.]
This implements a new internal client function which takes a revision
range and a (rev, path) pair as input (the "peg" for defining a line
of history), and returns new (rev, path) pairs at either end of the
range.  If the range extends into the future, then a sanity check is
done to guarantee all three paths lie on a single line of history.

As a proof of concept, 'svn cat -rN wcpath' now uses this function.
Regression tests to follow.

* subversion/include/svn_error_codes.h 
  (SVN_ERR_CLIENT_UNRELATED_RESOURCES):  new errorcode.

* subversion/libsvn_client/client.h
  (svn_client__prev_log_path, svn_client__repos_locations): new declarations.

* subversion/libsvn_client/ra.c
  (struct log_message_baton, svn_client__prev_log_path, log_receiver,
   svn_client__repos_locations):  new functions.

* subversion/libsvn_client/cat.c 
  (svn_client_cat): make 'svn cat' use the new history-finding
  function whenever invoked on a wc-path, but never on a URL.