CAN-2004-0749: path-based authz fixes for svn 1.1.  Patch 3 of 6.
Modify svn_repos_get_file_revs() to take an auth func and auth baton.

As soon as svn_repos_get_file_revs encounters a file that the user is
not authorized to see, stop tracing back in history and return only
the revisions discovered so far.

(Note: this patch assumes you've applied the 'common' patch first.)

* include/svn_repos.h (svn_repos_get_file_revs): Add
  authz_read_func and authz_read_baton args.

* libsvn_repos/rev_hunt.c (svn_repos_get_file_revs): Add
  authz_read_func and authz_read_baton args.  Stop tracing when an
  unauthorized path is encountered.

* libsvn_ra_local/ra_plugin.c (svn_ra_local__get_file_revs): Update
  caller to pass NULL authz func and baton.

* svnserve/serve.c (get_file_revs): Update caller to pass NULL authz
  func and baton.

* mod_dav_svn/file_revs.c (dav_svn__file_revs_report): Build an authz
    baton.  Invoke svn_repos_get_file_revs with the authz func and the
    baton.