Fix the horrible ra_dav->get_dir performance introduced in 1.1.0-rc4.
    ...also to be known as...

Fix the horrible ls performance over ra_dav introduced in 1.1.0-rc4.

If svn_repos_fs_revision_prop is supplied a path read authz function,
it will run that function on all changed paths for the revision to
verify that the revision is readable.  That is a very expensive
operation, but the best we can do when we simply want to read a
revision property.

The liveprops code in ra_dav already has a path and wants to read a
couple of revision properties for the last changed revision of those
paths.  By letting mod_dav_svn checking the path first, and not
letting svn_repos_fs_revision_prop do any authz we save an enormous
amount of work.

One test case (which I'm too lazy to describe how to set up here, but
the client operation was a non-recursive ls of a directory with 412
files) went down from six minutes to three seconds with this change.

* subversion/mod_dav_svn/liveprops.c
  (dav_svn_insert_prop, dav_svn_get_last_modified_time): Call
  dav_svn_authz_read on our path, and pass NULL as path read authz
  callback to svn_repos_fs_revision_prop.