* STATUS: Vote for r11172.
I've reviewed the change in its source code and tried patching it to the
branch under my Japanese environment. Since I am a partial committer for a po
file, which I think is involved in localization, I cast this vote.
If my vote is helpful, it will be nice. If it is not appropriate and is
unallowable, I'm sorry and please restore this vote.