Fix a compiler error on IRIX (Issue #2082).
N_(x) was improperly defined as (x) which creates problems with using it in the
context of a character array initilization.  The C standard doesn't allow
parenthesis around a string constant being used to initialize a character array.

It appears that the gettext people have run into this isuse and updated their
recommendation to using #define N_(x) x.  

* configure.in
* subversion/svn_private_config.hw
  (N_): Remove an invalid set of parens in the macro.