Merge r11486 from trunk to 1.1.x branch.
Approved by: +1: breser, ghudson, lundblad

Fix a compiler error on IRIX (Issue #2082).

N_(x) was improperly defined as (x) which creates problems with using it in the
context of a character array initilization.  The C standard doesn't allow
parenthesis around a string constant being used to initialize a character array.