Issue #1550: 'svn log' is broken for multiple WC paths.
Adjust the docstring for svn_client_log2 to match reality in that not
all combinations of URLs and working copies are permitted inputs.  In
the command-line client, throw an error if the user attempts to pass
some invalid combination of URLs and WC paths.  Since none of this
works currently anyway, it doesn't hurt to throw a sensible error
rather than giving odd output.  Improving 'svn log' to handle these
cases is left as a future enhancement.

* subversion/include/svn_client.h
  (svn_client_log2): This can either accept a URL followed by some
    relative paths or one WC path, nothing else.  Adjust the docstring
    to match the implementation.

* subversion/clients/cmdline/log-cmd.c
  (svn_cl__log): Verify that if a URL is the first target, no other
    URLs are present, and if a WC path is the first target, no other
    targets are given at all.  In either case throw a descriptive
    error message.