Kiss libsvn_fs_base node_revision_t caching goodbye.
The benefit of this cache was questionable once we no longer kept file
contents in the `nodes' table, and the complexity tradeoff when
dealing with trails and pools and such is just not worth the little
bit of benefit.  For example, during a checkout of, say, the
Subversion source code, the cache miss rate was *over 99%*.  Commit type
operations fared better, but I'm willing to take a little performance
hit there to reduce code complexity.

* subversion/libsvn_fs_base/dag.h
  (svn_fs_base__dag_get_node): Formatting change only (for internal
    consistency).

* subversion/libsvn_fs_base/dag.c
  (struct dag_node_t): Remove 'node_revision' member.  Also, drop a
    note about what is and isn't safe to keep in this structure.
  (cache_node_revision, uncache_node_revision): Remove.
  (set_node_revision, get_node_revision): Remove.  Callers changed to
    use the BDB layer functions directly, and (in some cases) to avoid
    unnecessary duplication because of now-matching pool lifetimes.
  (svn_fs_base__dag_dup): Lose meaningless comment.
  (svn_fs_base__dag_is_ancestor, svn_fs_base__dag_is_parent): Minor
    formatting changes.