Added partial French translation.
The translation focuses on 'help' messages.

It is a feature not to translate everything at the time.
The rationale behind is that an unclear backend-related English error
message approximated in French won't help a lot.

Some preliminary review by Marcel Gosselin and others.

SVN properties eol-style and mime-type set as other files around.

* subversion/po/fr.po: New file.