Make the regression test suite immune to issue #767, by doing merges from within the svn-test-work directory.
When running the tests in a ramdisk mounted at svn-test-work, merges
would fail because svn_io_file_rename() cannot not do a cross-device
move from the current directory -- the parent of svn-test-work -- into
some subdir underneath svn-test-work.

This workaround makes both the source and dest of the rename be
underneath svn-test-work, so there's no cross-device problem.  See
http://subversion.tigris.org/issues/show_bug.cgi?id=767#desc20 for
discussion of how to solve this in Subversion itself.

* subversion/tests/clients/cmdline/svntest/wc.py
  (State.copy): Add 'new_root' parameter to allow the copy to have a
  different wc root.

* subversion/tests/clients/cmdline/merge_tests.py
  (merge_similar_unrelated_trees,
   dry_run_adds_file_with_prop,
   merge_in_new_file_and_diff,
   merge_funny_chars_on_path,
   merge_keyword_expansions,
   merge_skips_obstructions,
   merge_binary_file,
   add_with_history): Use cwd kluge against potential cross-device problem.