Merged r15679 from trunk into the server-l10n branch.
A new module, "nls", was introduced into the trunk by Kouhei Sutou
since the last merge point.  This module is a subset of the
functionality I'd already factored out into the "intl" module.  For
now, the "intl" module will delegate to the "nls" module for
overlapping functionality.  In the end, they will likely need to be
merged into a single module.