Continue adding high-level logging to mod_dav_svn.  Followup to r15915.
Log 'svn ls URL'.

(Note that we're not logging GET or depth-0 PROPFIND requests, because
(1) apache's accesslog already logs them, and (2) we don't want to
fill the high-level log with the noise resulting from a bazillion 'svn
diff/merge' GET/PROPFIND requests.  Our current story is that GET and
PROPFIND are low-level things, so they'll continue to live in the
standard 'accesslog', rather than the high-level log.  This means 'svn
info URL' and 'svn cat' are considered low-level operations, not
high-level ones.)

* subversion/mod_dav_svn/repos.c 
  (dav_svn_do_walk):  if fetching dirents, set SVN-ACTION to "listing
                      entries of directory '/path'".