Teach mod_dav_svn to list SVNParentPath contents in a web browser.
This is controlled by a new 'SVNListParentPath' directive which, for
obvious security reasons, is 'off' by default.  If 'on', then only
child directories of SVNParentPath (presumably repositories) will be
listed in the html (or xml) response.

Note that this feature only works with http GET, not with 'svn ls'
(PROPFIND), nor with svnserve.  Those might be future improvements.

This change passes 'make davcheck' with the directive both 'on' and 'off'.

* subversion/mod_dav_svn/dav_svn.h
  (dav_svn_private_restype):  add new 'parentpath' private resource type.
  (dav_svn_get_list_parentpath_flag):  declare new func.

* subversion/mod_dav_svn/mod_dav_svn.c 
  (dav_svn_dir_conf):  new 'list_parentpath' boolean.
  (dav_svn_merge_dir_config):  inherit the boolean.
  (dav_svn_list_parentpath_cmd):  new function.
  (dav_svn_get_list_parentpath_flag):  new function.
  (dav_svn_cmds):  add new 'SVNListParentPath' directive.

* subversion/mod_dav_svn/repos.c   
  (get_parentpath_resource):  new helper func to build new private resource.
  (dav_svn_get_resource): look for an incoming SVNParentPath, and then
                          call helper function.
  (dav_svn_deliver): notice the new type of private resource, and list
                      (some of) the contents of the *real* directory.