Add support for creating XML schemas in various formats for the command-line client output to the UNIX build system.
NOTE: Currently, the configure script only checks for trang (the schema
conversion tool) as an executable, and not as a Java class run by the Java
interpreter.  We might find a way to extend this checking if we desire, but
in the mean time, the --with-trang configure option can be used.

* configure.in: Check for the trang tool, and add a new option --with-trang.
* Makefile.in
  (SCHEMA_DIR, TRANG): New variables.
  (mkdir-init): Ensure that the schema dir is created.
  (SCHEMAS_DTD, SCHEMAS_RNG, SCHEMAS_XSD): New variables.
  (schema, schema-rng, schema-dtd, schema-xsd, schema-clean): New rules.
  (.SUFFIXES): Add new schema file extensions.
  (.rnc.rng, .rnc.dtd, .rnc.xsd): New implicit rules.

* subversion/clients/cmdline/schema: Set svn:ignore property to ignore
  generated RNG, DTD and XSD files.