Followup to r15061: stop ignoring errors returned by svn_wc_prop_list().
r15061 changed svn_wc_prop_list() so that it no longer returns an error if
the requested properties file does not exist.

Two of the existing callers of svn_wc_prop_list() ignore all errors returned
from that function because they want to ignore the case where the path no
longer exists in the working copy.  As of r15061, this is no longer necessary.

* subversion/libsvn_wc/diff.c
  (load_base_props, change_dir_prop): Don't unconditionally ignore errors
    returned from svn_wc_prop_list(), since a not-found condition no longer
    returns an error.