Merge the svn_ra_replay implementation and associated tools from the ra-replay-dev branch into trunk.
In summary, this creates a new RA API that allows you to replay the
editor commands that were used to create a particular revision in a
repository.  This also includes some extensions to the underlying
replay API in libsvn_repos, to make it take into account authz, and
a simple replication program (svnsync) that uses the replay API to
copy the contents of one repository into another.

For the details of the changes in question, please see the logs from
http://svn.collab.net/repos/svn/branches/ra-replay-dev, r17440:17814.