Work on fixing diff_tests 32 (repos-wc diff showing added entries with props).
When performing a repos->wc diff, if the BASE->repos diff reported by
the server indicates that a file has been deleted (and so therefore we
must report the addition of a file currently in the wc), reconstruct
the propchanges necessary to create the current WORKING properties,
and report those propchanges in the file_added() callback.

This is incorrect for repos->BASE diffs, since the properties used there
should obviously be the BASE properties.  However, the current code also
needs to be changed to report the BASE file contents and MIME type,
and so is better handled in a later change.  (Additionally, handling
of _directories_ reported by the server as deleted during repos->BASE
diffs is completely broken; both features are tested in diff_tests 33).

Also note that the current handling of deleted files is wrong for
WORKING->repos file diffs (which, as we cannot currently test this
condition through the command-line tools, is also the case for much of
the 'svn diff' functionality).

* subversion/libsvn_wc/diff.c
  (delete_entry): Before reporting an added file (for a repos->wc diff),
    fetch the current WORKING properties and determine the changes
    necessary to create them from scratch.  Report the changes (together
    with an empty property hash) during the svn_wc_diff_callbacks2_t
    file_added() callback.  Also add two notes about inadequacies of
    the current implementation, described above, and rearrange comments
    slightly.