Continue work on fixing diff_tests 32 (repos-wc diff showing added entries with props).
When performing a repos->wc diff, if the BASE->repos diff reported by the
server indicates that a directory has been deleted (and so therefore we
must report the addition of a directory currently in the wc), reconstruct
the propchanges necessary to create the current WORKING properties,
and pass them to the dir_props_changed() callback.

Like r18098, this is incorrect for repos->BASE diffs, but the whole
handling of directories reported as deleted during repos->BASE diffs is
currently disabled due to an earlier check in directory_elements_diff()
(the function responsible for reporting both local BASE->WORKING diffs
and for faking added directories for repos->WORKING diffs).

[The diff is easier to follow with 'diff -w'; we're just adding a new
case before the 'if (modified)' condition, not changing it.]

* subversion/libsvn_wc/diff.c
  (directory_elements_diff): Before checking whether a directory has local
    property modifications, first check whether we are faking the addition
    of the directory (true only if we're doing a repos->WORKING diff,
    and the directory doesn't exist in the repos).  If so, fetch the
    current WORKING properties and determine the changes necessary to
    create them from scratch.  Report the changes (together with an empty
    property hash) during the svn_wc_diff_callbacks2_t dir_props_changed()
    callback.