Accept unambiguous substrings of a repository-relative path as head arguments.
When merging across multiple branches, it is often necessary to specify on
which of the multiple heads the operation must be performed (through
-S/--head). Until today, this parameter accepted either a working copy
directory or a full URL. This is unfortunate since the user really only wants
to specify one of a few possible heads.

This patch makes svnmerge accepts unambiguous substrings (of any length) of
one of the repository-relative paths recorded as head. For instance, it is
now possible to say: "svnmerge.py avail -S trunk".

Notice that this does not help in the case of "svnmerge.py init", because
you will probably need to specify a new head (as a full URL), not a previously
initialized one.

 * contrib/client-side/svnmerge.py:
   (main): Check if the head argument is a unambiguous substring of one the
   paths in the branch properties.
   (command_opts, OptionArg("-S")): modify the help string to specify that
   a substring is allowed.

 * contrib/client-side/svnmerge_test.py:
   (TestCase_TestRepo.testBidirectionalMergesMultiBranch): Use substrings
   for head parameters so to check the new feature.