Fix issue 1914 (update runs diff3 twice) by eliminating the second one.
In r9868 jpieper moved log file processing from close_file to
close_directory, to be more efficient at rewriting the entries
file. To be able to do per-file progress reporting, he added a
dry-run merge at close_file time (in merge_file).

With this change the log no longer contains a 'merge' command,
but rather only the commands to install the merge results in
the working copy.  So, we're still able to do per-file progress
reports and we'll still do per-directory entries updates.


* subversion/libsvn_wc/update_editor.c
 (merge_file): Take advantage of recently added
  svn_wc__merge_internal to create a log
  which updates the working copy with the
  merge results.