Fix issue #2613, svnsync does not mirror empty revisions exactly.
The problem was that if the source revision has no svn:author revprop
we would end up leaving the svn:author revprop from our initial commit.
The solution is to keep track of the revprops that were already there,
removing them from a hash as we copy the ones from the source.  If any
are left once we're done, those need to be removed.

* subversion/svnsync/main.c
  (copy_revprops): Add a new sync parameter, which makes us remove
   revprops in the destination revision that don't exist in the source
   revision.
  (do_initialize): Pass FALSE as the sync argument of copy_revprops.
  (do_synchronize): Pass TRUE as the sync argument of copy_revprops.
  (do_copy_revprops): Pass FALSE as the sync argument of copy_revprops.

* subversion/tests/cmdline/svnsync_tests.py
  (no_author): New test.
  (test_list): Add new test.

* subversion/tests/cmdline/svnsync_tests_data/no-author.dump: New test data.