Further followups to r22727.
Take advantage of the fact that a stream can now own a file handle,
allowing us to remove an explicit close of the new text base file, and,
more importantly, to ensure that the file is always closed before we
attempt to remove it (in the error case).

Also add a comment and remove two redundant error checks.

* subversion/libsvn_wc/adm_crawler.c
  (svn_wc_transmit_text_deltas2): Use non-deprecated
    svn_stream_from_aprfile2() so that the stream owns the new text base
    file.  Remove two redundant checks of 'err'.