Add a new runtime configuration operation ("mime-types-file") which, when present, specifies the location of a MIME types file (such as /etc/mime.types) which is used by the 'add' and 'import' operations for the auto-detection of MIME types of added/imported files.
I'm not especially thrilled with the change to the svn_client_ctx_t.
I don't want to parse the mime.types file multiple times in a single
invocation.  And, ideally, I wouldn't parse it for operations that I
know won't use it.  So it seemed natural to hang the types map on the
svn_client_ctx_t, but leave it NULL for everything except 'add' and
'import'.  However, this leaves the public API without any
cross-reference between this structure and the svn_client_add3() and
svn_client_import2() functions that use it.  But then again, you can't
tell from the public API that those things make use of the runtime
config's autoprop stuff, either.  Comments and suggestions welcome.

* subversion/include/svn_io.h,
* subversion/libsvn_subr/io.c
  (svn_io_parse_mimetypes_file): New.
  (svn_io_detect_mimetype2): Was svn_io_detect_mimetype.  Now accepts
    an optional mimetype_map parameter from which mime types are now
    chosen by default.
  (svn_io_detect_mimetype): Now just a wrapper around svn_io_detect_mimetype2.

* subversion/include/svn_config.h
  (SVN_CONFIG_OPTION_MIMETYPES_FILE): New.

* subversion/libsvn_subr/config_file.c
  (svn_config_ensure): Note the new mime-types-file option in the
    default 'config' configuration file contents.

* subversion/include/svn_client.h
  (svn_client_ctx_t): Add 'mimetypes_map' member.

* subversion/libsvn_client/add.c
  (svn_client__get_auto_props): Use svn_io_detect_mimetype2() instead of
    svn_io_detect_mimetype().

* subversion/svn/main.c
  (main): For 'add' and 'import', parse any configured MIME types file.