Add gen_method_defs, which helps automate the process of creating method definitions for the Perl bindings.
gen_method_defs parses the XML generated by Doxygen (if GENERATE_XML is
enabled in the doxygen.conf file) and uses it to generate a list of
methods with argument parameter definitions.  It doesn't fully automate
the process, as the generated content must still be reviewed, but it
does help with the initial boot strapping.

* subversion/bindings/swig/perl/native/gen_method_defs: Initial addition.