When harvesting copy committables, extend the use of the copy_pair loop iterpool to the actual call to harvest_committables(), and document why this is possible.
The primary concern is making sure the commit_items added to the committables
hash are not allocated from the loop pool.  But, because add_committable() uses
the hash pool to allocate the new commit_items, we can safely use the loop pool
when calling harvest_committables().

This is a followup to r23403.

* subversion/libsvn_client/commit_util.c
  (add_committable, harvest_committables): Update comments to note the fact
  that the new committable is allocated from the same pool as the committables
  hash.
  (svn_client__get_copy_committables): Use the iterpool for calls to 
  harvest_committables(), and add a comment explaining why this is possible.