Add support for revision range svnsync copy-revprops operations.
Prior to this change, using 'svnsync copy-revprops DEST_URL' would
throw an error because no revision-whose-props-to-copy parameter was
supplied.  Now, instead of throwing an error, the program will start
recopying the properties for all revisions previously synchronized.
If a single revision is supplied (which was required by the previous
syntax), the behavior is the same as it was -- only the properties for
that one revision are copied.  If, however, a revision *range* is
specified (M:N), then that range of revision properties is copied.

* subversion/svnsync/main.c
  (svnsync_cmd_table): Update the usage message for copy-revprops to
    reflect the new functionality, and tweak other usage messages for
    clarity.
  (copy_revprops_baton_t): Lose 'rev' for 'start_rev' and 'end_rev'.
  (do_copy_revprops): Handle revprop copies across ranges, now.
  (copy_revprops_cmd): Parse the now-optional revision now-range.