Avoid a case of potential working copy corruption when an update/switch adds a file that was already scheduled for addition.
NOTE: This is a marginal backwards incompatible format change to the
working copy because of the addition of the force attribute to the
modify-entry log command.
If a new client uses this attribute (for example in the case mentioned
above), something fails and an old client is used for the cleanup, then
this flag will be ignored by the old client.  This will not result
in a corrupt working copy, but will leave the added file in the schedule add
state.  I consider this marginal case not worth a format upgrade by itself.

* subversion/libsvn_wc/update_editor.c
  (merge_file): For an added file that was scheduled for addition, tweak
  the schedule field of the file's entry loggily.  In the same case
  don't install an empty "fake" text base for merging, but instead
  create a temporary empty file as the left side of the merge.

* subversion/libsvn_wc/log.c
  (SVN_WC__LOG_ATTR_FORCE): New macro.
  (log_do_modify_entry): Take the force flag in the log entry into account
  when modifying the entry.
  (svn_wc__loggy_entry_modify): Set force attribute in the log
  entry if the corresponding modify flag is set.