JavaHL: More work on the info() and info2() interface.  Followup to r24274.
This commit removes duplicate implementations of info() and info2() at the C++
layer in favor of a thin wrapper at the Java layer.  Because info() has a flags
from the wc entry which info2() doesn't, we still get to fetch the entry in the
receiver.  In the case of the receiver being called by a repository crawl, we
also have to make sure we cache the wc root path beforehand, so we can
reconstruct the absolute path to fetch the entry correctly.

The alternative to this convoluted double-entry fetching is to add a couple of
fields to svn_info_t, but this impacts *every* call to svn_client_info(), not
just when using JavaHL.

(Note: The approach proved fruitless, and most of the work done here was
reverted in r24531.)

[ in subversion/bindings/java/javahl/ ]
* src/org/tigris/subversion/javahl/SVNClient.java
  (info): Reimplement as a wrapper around info2().

* native/SVNClient.h,
  native/SVNClient.cpp
  (info_baton): Add a wcPath member.
  (info2, info, info): Consolidate the info implementations into a single
  function.
  (infoReceiver): Update the receiver to use store the information retrieved
  locally from the wc entry.

* native/org_tigris_subversion_javahl_SVNClient.cpp
  (Java_org_tigris_subversion_javahl_SVNClient_info):  Remove.
  (Java_org_tigris_subversion_javahl_SVNClient_info2):  Use the new combined
  info() implementation.