Finish issue #2474 - Change conflict file naming scheme.
As of this commit, the format chosen for the conflict files is as follows:

   FILENAME.rREV[.FILEEXT]
   FILENAME.mine[.FILEEXT]

In other words, for extensionless files, there is no change at all.
But for files which have an extension (.doc, .txt, etc.), the
extension is repeated after the ".mine" or ".rREV" revision
indicators.

(Oddly, 'make check' passes without test changes, probably because the
files in our "Greek tree" don't have extensions.)

* subversion/libsvn_wc/update_editor.c
  (merge_file): If the versioned file has an extension (as determined
    by svn_path_splitext()), append that extension to the end of the
    conflict file names.