Test how svn_path_splitext() handles dotfiles, too.
In IRC, eh, maxb and I agreed that the file ".foo" has no extension,
and the file ".foo.txt" has an extension of "txt".  However, this is
not how Python's os.path.splitext() handles said files, and so is not
the way I originally wrote svn_path_splitext().  So, expect this test
to fail now until I fix svn_path_splitext().

* subversion/tests/libsvn_subr/path-test.c
  (test_splitext): Test the handling of dotfiles, too.
  (test_funcs): Expect test_splitext() to fail now.