Ignore branching revisions when running 'svn log -g'.
To do this, we basically duplicate the implicit mergeinfo calculation from
libsvn_client/copy.c:get_implied_mergeinfo(), and check it against the current
mergeinfo to "guess" that the path and revision in question is a branching
revision, and ignore it for generating the mergeinfo diffs.

Because of layering issues, and circular dependencies, this is implemented as a
filter callback function passed to svn_fs_mergeinfo_for_tree().

Note:  This fixes one problem at the expense of creating a less serious one,
hence the marking of log_tests 19 as XFail.  The cause is known and a solution
under investigation.  (This was fixed in r25594.)


* subversion/libsvn_fs_util/mergeinfo-sqlite-index.c
  (get_mergeinfo_for_children, svn_fs_mergeinfo__get_mergeinfo_for_tree):
  Add filter_func arguments and filter each path mergeinfo is found on.

* subversion/include/svn_fs.h
  (svn_fs_mergeinfo_filter_func_t): New interface.
  (svn_fs_get_mergeinfo_for_tree): Accept filter_func and filter_func_baton
  parameters, and document their use.

* subversion/include/private/svn_fs_mergeinfo.h
  (svn_fs_mergeinfo__get_mergeinfo_for_tree): Add filter_func and
  filter_func_baton parameters.

* subversion/libsvn_fs/fs-loader.c
  (svn_fs_get_mergeinfo_for_tree): Add and propogate filter_func and
  filter_func_baton parameters.

* subverion/tests/cmdline/log_tests.py
  (test_list): Unmark merge_sensitive_log_branching_revision() as XFail,
  and mark merge_sensitive_log_branching_revision() as XFail.

* subversion/libsvn_repos/log.c
  (filter_baton): New baton structure.
  (calculate_branching_copy_mergeinfo): New function.
  (branching_copy_filter): New function.
  (get_combined_mergeinfo): Use branching_copy_filter() to remove extra
  mergeinfo when search for child log messages.