Improve performance by avoiding redundant server round-trips when using -S.
This patch introduces an internal cache in get_repo_root() which
allows the function to automatically detect if a given URL is part
of an already analyzed repository without doing a server roundtrip.

Since I was at it, I also removed a useless global declaration for
the other similar cache used by get_svninfo().

* svnmerge.py
  (get_svninfo): Remove global declaration.
  (get_repo_root): Update and use the new _cache_reporoot,
    storing repository roots that were already found.

* svnmerge_test.py
  (TestCase_SvnMerge.svnmerge2): empty the new cache.