Add a depth argument to svn_repos_begin_report2, and get rid of svn_repos__begin_report.
Background: a while ago we decided that svn_repos_begin_report2 doesn't need
to take depth, because we also pass depth to svn_repos_set_path/link_path.
Later we realized that in order to implement depth upgrades we do need a depth
argument, but we didn't add it to svn_repos_begin_report2.  Instead we used
svn_repos__begin_report, which was originally created precisely because
_begin_report2 didn't take depth, and is now obsolete.

* subversion/include/private/svn_repos_private.h:
  Delete (svn_repos__begin_report was the only function in this file).

* subversion/include/svn_repos.h
  (svn_repos_begin_report2): Add and document a depth argument.
  (svn_repos_begin_report): Tweak docstring.

* subversion/libsvn_repos/reporter.c
  Don't include private/svn_repos_private.h.
  (svn_repos__begin_report): Remove.
  (svn_repos_begin_report2): Add a depth argument and move the body of
   svn_repos__begin_report here.
  (svn_repos_begin_report): Call svn_repos_begin_report2 instead of
   svn_repos__begin_report.

* subversion/mod_dav_svn/reports/update.c
* subversion/svnserve/serve.c
* subversion/libsvn_ra_local/ra_plugin.c
  Don't include private/svn_repos_private.h, and call svn_repos_begin_report2
  instead of svn_repos__begin_report.