Add a script to generate a Java class which provides a mapping from C error codes to a Java enum.
This commit also addes the generated Java class to the repository.  Normally,
we don't version generated files, but I think this case warrants an exception.
The C error codes upon which the generated file is based don't change very
often, and futzing with the build system to autogenerate the class would be
Hard (given the need to build and install the python bindings first).  Manually
updating the ErrorCodes.java file when the error codes change should be
sufficient.

* tools/dev/gen-javahl-errors.py:
  New

* subversion/bindings/javahl/src/org/tigris/subversion/javahl/ErrorCodes.java:
  New, generated from tools/dev/gen-javahl-errors.py