JavaHL: Remove redundant docstrings in implementors of the SVNClientInterface.
When a class implements an interface, Javadoc automagically inherits the
description, @returns, @param, and @throws tags from the interface so, we do
not need to duplicate that bit of data in the derived classes.  This commit
removes the redundant documentation.

This also deprecates old versions of methods, with the recommendation that
comsumers use the latest API.  These methods continue to function, of course.

[ in subversion/bindings/javahl/org/tigris/subversion/javahl/ ]
* SVNClient.java,
  SVNClientSynchronized,
  SVNClientInterface:
  Remove and update docstrings as described above.