Merge sqlite-node-origins branch to trunk.
There is now a sqlite table for node_origins in the same database file
as mergeinfo (though perhaps it should be moved for concurrency
reasons) and support code in libsvn_fs_util.  FSFS uses it (BDB uses a
BDB table); it is currently only filled at CACHE MISS time, though it
should be straightforward to fill it at commit time as well.