Remove svn_depth_exclude from the API.  It was an implementation detail of some of the modules which processed depth, not a semantically meaningful depth itself.
Arguably, it *would* be useful to have a concept of svn_depth_exclude,
so that you could do things like "check all of A except for A/D".
However, we did not implement that (at least, we didn't document it if
we did, though it may have worked by accident).  Perhaps we'll add
this back later, completely supported.

* subversion/include/svn_types.h
  (svn_depth_exclude): Remove.

* subversion/libsvn_subr/kitchensink.c
  (svn_depth_to_word, svn_depth_from_word): Remove "exclude" cases.