Remove the almost-completely-unused svn_fs_change_mergeinfo API.
Why "almost"?  Well, there's a C test that uses it, but that C test is
pretty wacky already (the mergeinfo it sets is invalid, it does a
svn_fs_get_mergeinfo on a non-existing path, etc).  And it doesn't
actually verify anything about the return value from
svn_fs_get_mergeinfo.  So ditch the test too.

(This change is similar to r28146 on the
sqlite-mergeinfo-without-mergeinfo branch; I should have done it on
trunk in the first place.)

* subversion/include/svn_fs.h
  (svn_fs_change_mergeinfo): Remove.

* subversion/libsvn_fs/fs-loader.c
  (svn_fs_change_mergeinfo): Remove.

* subversion/libsvn_fs/fs-loader.h
  (root_vtable_t): Remove change_mergeinfo.

* subversion/libsvn_fs_base/tree.c
  (change_mergeinfo_args, txn_body_change_mergeinfo,
   base_change_mergeinfo): Remove.
  (root_vtable): Remove base_change_mergeinfo.

* subversion/libsvn_fs_fs/tree.c
  (fs_change_mergeinfo): Remove.
  (root_vtable): Remove fs_change_mergeinfo.

* subversion/tests/libsvn_fs/fs-test.c
  (get_mergeinfo): Remove.
  (test_funcs): Remove get_mergeinfo.