Fix bug in merge notifications involving deletions/replacements.
Previously, when a merge deleted a path the deletion notification(s) came
before the "--- Merging*" | "--- Reverse-merging*" headers.  This change
ensures the header always shows first.

A beneficial side-effect of this change is that merge delete and replacement
notifications are now akin to those from update -- Specifically, if a
directory is deleted we no longer report on every child under that directory
as deleted too.

* subversion/libsvn_client/merge.c
  (merge_dir_deleted): Pass NULL notify function/baton to
  svn_client__wc_delete(), instead let repos_diff.c:delete_entry() do the
  notifications.

* subversion/tests/cmdline/merge_tests.py
  (delete_file_and_dir, merge_dir_replace):
* subversion/tests/cmdline/revert_tests.py
  (status_of_missing_dir_after_revert_replaced_with_history_dir):
  Tweak expected merge output to reflect changes above.