Merge the correct-copy-mergeinfo branch (r28402:28462) to trunk.
Now only WC->WC copies/moves create empty mergeinfo on the copy
destination if no explicit mergeinfo existed on the source.  The three other
copy varieties (WC->URL, URL->WC, URL->URL) *can* result in empty mergeinfo
on the destination, but only if it the source explicitly had it to start
with.