For issue #3089: cause mod_dav_svn to report an error more recognizeably.
There is some question as to whether this is the right solution, see:

   http://subversion.tigris.org/servlets/ReadMsg?list=dev&msgNo=135089
   From: Karl Fogel <kfogel@red-bean.com>
   To: dev@subversion.tigris.org
   Subject: Better error reporting from mod_dav_svn?
   Date: Thu, 14 Feb 2008 00:36:40 -0500
   Message-Id: <87r6fgumjb.fsf@red-bean.com>

We may stick with this method, or we may switch to some other method
of letting the client detect the "mergeinfo is unimplemented" error.
Either way, it is independent of other changes related to issue #3089,
so I've extracted it from r29314 on the mergeinfo-capability branch.
See the r29358 log message for more about how changes from that branch
are being ported to trunk.

[Note: changed to a more conservative solution for 1.5, see r29366.]

* subversion/mod_dav_svn/reports/mergeinfo.c
  (dav_svn__get_mergeinfo_report): Don't flush before returning error,
    as flushing might obscure the error due to certain intricacies of
    mod_dav's plumbing.