The reintegrate code tries to make sure that every subtree in the source has the "same amount" of the target merged to it.  This change fixes a bug where a path which has *none* of the target merged to it is ignored.
Note, though, that while this is correct in general, files *added* on
the branch (or renamed, etc) probably need to be handled specially.
Because they aren't (and because the only reintegrate tests do test
the rename-on-a-branch scenario), all of the real reintegrate tests
are now XFail.

See a thread starting here for more details:
   http://svn.haxx.se/dev/archive-2008-02/0556.shtml

* subversion/libsvn_client/merge.c
  (remove_irrelevant_ranges): Rename new_by_path to new_catalog, and
   fix its type.  Remove redundant check that filtered_mergeinfo is
   not NULL, and keep its information around even if the mergeinfo is
   empty, since empty mergeinfo still overrides inheritance.

* subversion/tests/cmdline/merge_tests.py
  (test_list): XFail the two basic reintegrate tests.