Fix spurious Win32 buildbot failure of diff_tests.py 47
On Win32,if your current directory is not a working copy, and you do
something like

  diff -x WCPATH -r1

You'll get the error "svn: '.' is not a working copy" because WCPATH is
interpreted as the --extensions argument.  How other platforms get far 
enough to examine the --extensions arg and return the error
"svn: Invalid argument 'WCPATH' in diff options." is a mystery to me, 
but diff_tests.py 47 expects the latter and does pass on other platforms 
(per the buildbots).  Regardless, the former error is quite legitimate on 
Win32.

* subversion/tests/cmdline/diff_tests.py
  (diff_summarize_xml): Tweak expected error regexp to also accept the error 
  that is produced on Win32.