Don't waste CPU time creating deltas against the empty stream; just use a single insert operation.
The only reason to represent deltas against the empty stream as
anything other than a single insert operation would be to "compress"
it; but since most of the time we're gzipping the delta stream anyway,
this is usually a waste of time.

This does mean that more bandwidth will be used when talking to
pre-1.4 clients, and more disk space will be used when storing data in
pre-1.4.x FSFS repositories.  However, by the time 1.6.0 is released,
pre-1.4.x clients will be very old, and there is an "svnadmin upgrade"
command to upgrade old repositories.

* subversion/libsvn_delta/text_delta.c
  (compute_window): Replace call to svn_txdelta__vdelta with a single
   call to svn_txdelta__insert_op.

Patch by: Niels Werensteijn <n.werensteijn@student.utwente.nl>