Show context in the "conflict-only diff3" output.
The header line for conflict-only context sections looks like:

@@ <M,N |M,N >M,N @@

If there is prior art that I should be mimicking, let me know.

This revision implements it only for in-memory diffs (because it's
easier to start here...), so the only visible way to see that it works
is to see that the test XFails during the file-diff part, not the
mem-diff part.

* subversion/libsvn_diff/diff_memory.c
  (merge_output_baton_t): Add some fields only used by conflict-only
   mode: hunk, hunk_length, hunk_start, real_output_stream, and pool.
  (output_merge_token_range): Add a lines_printed_p arg.
  (output_common_modified, output_latest, output_conflict): Adjust.
  (output_conflicts_flush_hunk): New.
  (output_conflict_with_context): Actually have context.
  (svn_diff_mem_string_output_merge2): Initialize fields and clean up
   differently for conflict-only mode.

* subversion/tests/libsvn_diff/diff-diff3-test.c
  (test_three_way_merge_conflict_styles): Add context to expected
   output.
  (test_funcs): Mark test_three_way_merge_conflict_styles as XFail.