Merge the in-memory-cache branch to trunk.
See that branch's log for full details, but in short:

The branch rewrote various in-memory caches in FSFS to use a unified
API.  By default this API is implemented by some in-process hashes,
but via a configuration file (db/fsfs.conf), it can use memcached
instead.  Additionally, expanded fulltexts are cached if memcached is
available.

memcached is accessed with the apr_memcached API, which will be in a
future version (1.3.x branch, now) of apr-util, and is also available as an
separately-distributed library.  The build system knows how to deal
with both.

Other things that changed:

* The test suite now has a config file; currently, the config file is
  just used as the FSFS config file for any created repositories.

* New random APIs: svn_base64_encode_string2, svn_stream_from_string.

* ra_local ignores "FS warnings" instead of aborting.