Fix a test failure in Ruby bindings.
In r31211 on trunk our notifications were changed to send paths
differently.  In r31348 the Ruby tests were updated for the context
changes and in r31356 this was merged to the 1.5.x branch.  The test
contains 1-line of code that needs to be modified for the branch.

* subversion/bindings/swig/ruby/test/test_client.rb
  (test_mkdir_p): Update test for 1.5.x

Patch by: kou

Votes:
  +1: markphip, cmpilato