Fix issue #2939 over ra_neon:  bogus existence check.
See reproduction recipe in the issue to verify this fix -- it uses
svnmucc.  Scenario can't be tested using the normal 'svn' client,
since it involves doing a directory replacement within a single
transaction.  svnmucc and gvn can create the scenario, however,
resulting is a revision which (until now) couldn't be svnsync'd over
http!

* subversion/libsvn_ra_neon/commit.c (commit_add_file):
    do PROPFIND existence check on the *working* url (i.e. the path
    within the transaction), not the public url (which the path in HEAD).