In the Subversion libraries, replace uses of "assert" and "abort" with the new SVN_ERR_ASSERT and SVN_ERR_MALFUNCTION macros where possible. Replace "if (cond) abort();" with "SVN_ERR_ASSERT(!cond)".
This is a like-for-like change except that the new macros allow the
application at run time to specify an alternative behaviour such as
returning an error object instead of aborting, and they are not affected by
the NDEBUG macro (so they are not compiled out in release-mode builds as
"assert" could be).

Omit libsvn_ra_serf because it uses assert and abort as placeholders for
unfinished development.

(61 files affected.)