Finish issue #3220 ('svn log -g' seems quite slow) and, as a bonus side-effect, also finish issue #3235 ('log -g' merge determination logic faulty).
[Merge the entirety of the issue-3220-dev branch to trunk.]

Prior to this change, the method used by 'log -g' for determining what
merges were made in a revision was:

   - fetch all the mergeinfo found in REV
   - fetch all the mergeinfo found in REV-1
   - compare, contrast, and collapse the results

The problems with this approach are two-fold.  First, it's slow, and
it gets slower as more paths in a revision grow mergeinfo.  Secondly,
it's just wrong.  Just because mergeinfo appears or disappears doesn't
mean a merge was done -- perhaps a path that held some mergeinfo was
merely deleted.

This new approach is to look at only the paths changed in REV, and see
if/how their mergeinfo changed.  This is much, much faster on average
for a repository actually using mergeinfo, and much more accurate as
well.

* subversion/libsvn_repos/log.c
  (fs_mergeinfo_changed): New (pretty-generic) mergeinfo change
    calculation function.
  (get_combined_mergeinfo_changes): New, supercedes ...
  (get_combined_mergeinfo, get_merged_rev_mergeinfo): ...these, which
    are now removed.
  (do_logs): Use get_combined_mergeinfo_changes() instead of
    get_merged_rev_mergeinfo().

* subversion/libsvn_repos/repos.h
  (svn_repos__prev_location): New prototype.

* subversion/libsvn_repos/rev_hunt.c
  (svn_repos__prev_location): Was prev_location(), now module-public.
    Allow prev_rev, prev_path, and appeared_rev parameters to be NULL,
    in which case (for each) don't try to set their pointer values.
    Callers updated.