Fix an incompatibility with BSD make, which was introduced in r24293.
Even though r24293 was meant to fix a compilation failure in the
bindings when compiling Subversion from a different directory than
where the source tree is located, this still didn't work for me.

To illustrate, a command sequence like

  $ svn co https://svn.collab.net/repos/svn/trunk subversion
  $ mkdir obj
  $ cd obj
  $ ../subversion/configure --prefix= ...
  $ make
  $ make swig-py

failed with:
make: don't know how to make subversion/bindings/swig/python/svn_client.c. Stop

It turned out to be a problem specific to BSD make.
With GNU make, things worked fine.

The problem was that in build-outputs.mk, the generated binding
source files were preceeded with '$(top_builddir)/' when listed
as targets, but did not have any prefix (i.e. were relative paths)
when listed as dependencies. BSD make is apparently too naive to
figure out that e.g. 'foo.c' and './foo.c' are meant to refer to
the same file.

Since '$(top_builddir)' is hardcoded to just '.' (a dot)
at the top of Makefile.in, it is redundant and can be ommitted.
As soon as the target and dependency paths matched, the problem
went away. Turns out GNU make does not care either way, so this
commit makes both make implementations happy.

* build/generator/gen_make.py
  (Generator.write): Don't prefix the names of generated binding
    source files with '$(top_builddir)/' when listing them as target,
    because this confuses BSD make. When listed as dependencies,
    the files do not have this prefix either.