Include more information in the baton passed around during a merge. No functional change; this is a step in preparation for other changes.
During a merge, pass around in the merge baton the repository location of
the merge sources, instead of just the URL of the merge-right source, so
that the information is available to all callbacks. (Some of the callbacks
already receive some of this information through the svn_diff_callbacks3_t
interface, but inconsistently and incompletely. This change is a local
alternative to revving the callbacks API and making it consistent.)

* subversion/libsvn_client/merge.c
  (merge_source_t): Move ahead of merge_cmd_baton_t declaration.
  (merge_cmd_baton_t): Replace 'url' member with a merge_source_t member
    'merge_source'.
  (merge_file_added, merge_dir_added): Use new merge_cmd_baton_t member.
  (do_merge): Set merge_cmd_baton_t's new 'merge_source' member to reflect
    the current merge source being merged.