Merge the issue-2843 branch back to trunk.
I wasn't able to use reintegrate, see 
http://subversion.tigris.org/servlets/ReadMsg?list=dev&msgNo=145078.  So 
this was done the old-fashioned way, with a 2-URL merge:
http://svn.collab.net/repos/svn/trunk@34105
http://svn.collab.net/repos/svn/branches/issue-2843-dev@34105.

Note: This merge added mergeinfo on the subversion directory.  This
mergeinfo originated in r31940 on the issue-2843 branch, but there is no
obvious reason why it was there.  r31940 was a merge of r31359-31939 from
trunk, but there is no mergeinfo diff between those revs, so where this came
from is a complete mystery.  In that light I removed this mergeinfo prior to
the commit.