Make cmdline test update_tests.py 50 pass.
tree_conflicts_on_update_2_3 expected to get a "Skipped" message when
trying to update from *within* a directory deleted on the repository
(that also happens to be a tree-conflict victim). The tree-conflict-ness
does not change anything about a path missing in the repos, and
without stepping out to the parent dir, no tree-conflicts can be known
on the directory. See also this message:
Date: Wed, 26 Nov 2008 06:32:57 +0100
From: Neels J Hofmeyr <neels@elego.de>
Subject: update_tests.py 50 XFails.

* subversion/tests/cmdline/update_tests.py
  (tree_conflicts_on_update_2_3): 
    Don't expect to "Skip" when updating *within* a "deleted" dir, twice.
    Comment on that. Remove an obsolete warning comment.
  (test_list): Don't XFail tree_conflicts_on_update_2_3, it passes now.