* TODO-1.6: Remove the svn_dirent_*/svn_uri_* from the 1.6 blocker list
I reviewed the public api to make sure it is in a state that we can support
for 1.6+. 

This new API cleans up a lot of the old unified path code by separating code
paths for different path types and it already resolves several path 
normalization issues of previous releases. Most importantly it allows future
improvements of the separate codepaths which was almost impossible via the
one path api for all approach previously used.

See also
http://subversion.tigris.org/ds/viewMessage.do?dsForumId=462&dsMessageId=1106748