Deal with versioned files obstructed by versioned directories.
When we ask for the sqlite database for "/some/file", and it turns out
that somebody has moved a versioned directory onto "file", then our search
turns up "/some/file/.svn/wc.db" erroneously. Information on "file" is in
the "/some/.svn/wc.db" database.

To correct the problem, if it looks like the above situation is possible,
then we look in the parent database (if present) for information on
"file". If the entry *is* supposed to be a file rather than a subdir, then
we return that parent database instead.

Note that this can only occur when we have per-dir .svn subdirs. When
wc-ng is complete, this scenario will not be possible.

This fixes (at least) stat_tests.py 5.

* subversion/libsvn_wc/wc_db.c:
  (struct svn_wc__db_pdh_t): record if the directory actually represents
    an obstructed file.
  (enum statement_keys, statements): add two new statements to assist with
    looking for obstructed 'file' nodes.
  (parse_local_abspath): if the situation makes it possible for an
    obstructed file, then check it out and return the appropriate PDH.
    make a slight fix to how we store PDH->wcroot_abspath.