Tweak the logic that maintains the 'pending_error' RA session variable in libsvn_ra_serf.
libsvn_ra_serf's error handling is not straightforward, and is (in my
experience) unreliable.  The idea of using a session variable to hold
pending errors and maintaining that variable's state in multi-various
locations makes me nervous.  This commit adds some assertions to
ensure that we aren't overwriting errors (thus leaking them).  This
seems to have had the effect of causing a test that was expected to
fail to start passing, possibly because the only reason the test was
failing before this commit was due to memory leak detection logic
abort()s.

* subversion/libsvn_ra_serf/ra_serf.h
  (svn_ra_serf__handle_client_cert, svn_ra_serf__handle_client_cert_pw):
    Note that these function might set the session's 'pending_error' member.

* subversion/libsvn_ra_serf/util.c
  (svn_ra_serf__context_run_wait): Rather than clobber any pending
    error stored in the session (thereby leaking it, thereby causing
    an abort() in maintainer mode), assert that there is no error.
    If, after we do real work, the 'pending_error' variable is
    non-NULL, make sure we reset it before returning its value to the
    caller.
  (handle_response): Add some asserts to make error clobbering a more
    direct failure case.
  (svn_ra_serf__get_relative_path): Formatting tweak only.

* subversion/libsvn_ra_serf/update.c
  (finish_report): Reset 'pending_error' after destroying the error.
    Minor additional logic juggling nearby.

* subversion/libsvn_ra_serf/replay.c
  (svn_ra_serf__replay_range): Minor logic tweak.

* subversion/tests/cmdline/switch_tests.py
  (relocate_and_propset): Add comment about test failure in Neon.
  (test_list): Adjust XFail() for relocate_and_propset() test.