You'd think some sanity might exist SOMEWHERE in libsvn_wc. Apparently not, as this latest oddity really takes the cake.
Refine how we set the COPIED flag for schedule-delete nodes. We don't want
to set it if we have the changed_* information for a node. Why? Pfft. Why
ask why?

This keeps copy_tests 24 and schedule_tests 10 working (the driving force
behind the recent work here). But also fixes the became-broken tests of
update_tests 54 and switch_tests 31. Sigh.

* subversion/libsvn_wc/entries.c:
  (read_entries): if cmt_rev is valid (aka changed_rev in the new world),
    then we do NOT want to set the copied flag.