Fix up merge_tests 101. This has been failing due to API Errata wc003.txt.
In short, the merge code records state for all children in a copied tree
as if each child was individually added. However, the *net* effect is the
same as a single copy of the root. The wc-ng code coalesces these apparent
multiple adds into a single add of the root.

This has a visible effect of removing the 'A' status on child nodes, and
adding just the root (and recursively, its children) at commit time.

Adjust the test to reflect this New World Order.

* subversion/tests/cmdline/merge_tests.py:
  (commit_to_subtree_added_by_merge): adjust test output per wc003.txt