Stop propagation of self-referential mergeinfo via --reintegrate merges.
While we recommend that users delete a branch once they reintegrate it, it
should still be possible to keep sync merging to the branch if a --record-only
merge is performed to avoid cyclical merge problems.  However there is a bug
where, even if the record only merge is performed correctly, subsequent
reintegrations can produce self-referential mergeinfo on the target.  This
mergeinfo can then propagate to other branches where it can appear as
legitimate (non-self-referential) mergeinfo.  This in turn can make it
appear that sync merges have been performed when they have not and
vice-versa 

* subversion/libsvn_client/merge.c
  (merge_cmd_baton_t): New boolean flag reintegrate_merge.
  (filter_self_referential_mergeinfo): *Do* filter during reintegrate merges.
  (do_merge): Add boolean flag signaling reintegrate merges and use it to
  initialize the new merge_cmd_baton_t member.
  (merge_cousins_and_supplement_mergeinfo, svn_client_merge3,
  svn_client_merge_peg3): Update calls to do_merge(). 

* subversion/tests/cmdline/merge_tests.py
  (basic_reintegrate, reintegrate_with_rename,
  reintegrate_with_subtree_mergeinfo): Don't expect notifications from the
  application of self-referential mergeinfo during a reintegrate merge.
  (multiple_reintegrates_from_the_same_branch): New test showing how
  reintegrate can produce self-referential mergeinfo and the problems that
  can arise from it.
  (test_list): Add multiple_reintegrates_from_the_same_branch.