Fixed an issue where commit over ra_serf segfaulted.
Reason of the segfault is that an APR_EOF status code, indicating svn handled
the response to a message completely, was wrapped in an error object and 
handled as an error, while it's actually the desired result. This is a 
temporary fix to get commit working again. I guess the better fix would be 
to stop wrapping desired status code(s) in error objects and just 
returning APR_EOF to serf if svn handled the reponse completely.

* subversion/libsvn_ra_serf/commit.c
  (handle_checkout): treat an APR_EOF error as a correct result (but still
   return it to the caller).