Follow-up to r862679 and r878607.
'dav_svn__location_header_filter' is no-op prior to this commit, I mean
still clients originating the commit are seeing the Location response headers
with Master URI.

I believe this code was working well prior to r862679 where the code did the
opposite of what was there in the associated doc and the log.

In r878607 filter function was returning 'APR_SUCCESS' but that causes 
failure, reinstating the original 'return ap_pass_brigade(f->next, bb);' 
makes thing working.

This commit changes the Location header from the Master and changes the same
with the proxy server URI.

* subversion/mod_dav_svn/mirror.c
  (dav_svn__location_header_filter):
   Fix the logic to determine the subrequest.
   Return ap_pass_brigade(f->next, bb) instead of APR_SUCCESS.