Fix "command" return value under Windows.
For sys.platform == 'win32', the list of command arguments is turned
into a single string containing all of the arguments, quoted.  But
this command string was also returned by the method, to callers who
expected it to be a list of strings.  In the case of an error, this
return value is processed with ' '.join(), resulting in error messages
like "C : \ p a t h \ t o \ s v n . e x e   a r g 1   a r g 2".

This change returns the commands as a list of strings regardless of
whether the command list is turned into a string internally.

Patch by: James Abbatiello <abbeyj@gmail.com>

* subversion/tests/cmdline/svntest/main.py

  (open_pipe): Return the command list that was passed to the
  function, regardless of whether it needs to be converted to a string
  internally.