[[ Reverted in r39153, see discussion on list. * *  The possible solution for keeping backwards compatibility *  'making access baton close a no-op' doesn't work because *  in that case we can never close wc.db files before deleting *  a working copy with an independent access baton in the same *  parent pool. ]]
When opening access batons, recursively lock all its subdirectories to an
infinite depth as well.  In wc-ng, we don't have the notion of recursively
opening a database; it's either open or it's not.  By making this change now,
we can remove the existing complexity of determining how many levels to lock
and where to lock them.

We also gain another advantage.  When attempting to open a baton, we can check
and see if that baton is already opened, and just return it if it is.  By
always recursively locking, we can guarantee that a found baton is also
recursively locked, which is what the original caller may have intended.
All this simplifies our current access baton usage, and the path to removal.

Oh, and it incidentally fixes copy test 76.

* subversion/tests/cmdline/copy_tests.py
  (commit_copy_depth_empty): Remove comment.
  (test_list): Mark test as passing.

* subversion/libsvn_wc/lock.c
  (open_single): Check and see if a baton exists before opening a new one.
  (do_open): Always recursively lock to an infinite depth.  Also, don't
    duplicate the missing placeholder in the shared set.