Bump the working copy format to 13. The primary change is the introduction of a new WORK_QUEUE table, destined to replace the classic "loggy" infrastructure.
Note: this is a silent, automagic upgrade from f12 to f13. If you have a
working copy that is format 11 or older, then "svn upgrade" will be
required, as before.

Note: we were intending some schema changes to better handling conflicts,
but due to open questions, that has been postponed to an f14 bump.

* subversion/libsvn_wc/wc.h:
  (SVN_WC__VERSION): bump to 13

* subversion/tests/cmdline/upgrade_tests.py:
  (get_current_format): return 13, as the current format