Following up on r39678 and r39709, make the notifications for the accept_deleted nodes consistent.
Before r39678 all changes on these nodes below a tree conflict were
suppressed in the update editor. This is a cleanup change updating
the notification type for only these changes.

Note: I don't see a problem with making the svn client skip these
new notifications to give the same output as the 1.6 svn client, but I
would like to see these notifications when using the libsvn_client api.
This would allow reverting the test changes from this revision and r39678.

* subversion/include/svn_wc.h
  (svn_wc_notify_action_t): Add two notification types, moving all new in
    1.7 notifications regarding updates together.

* subversion/libsvn_wc/update_editor.c
  (file_baton): Remove unused boolean.
  (make_file_baton): Remove initializing of unused boolean.
  (do_entry_deletion): Notify deleted below accept_deleted.
  (add_directory): Use new notification for add below accept_deleted.
  (close_directory): Use new notification for updated below accept_deleted.
  (close_file): Don't check for tree conflicts, because these are already
      handled and check for accept_deleted.

* subversion/svn/notify.c
  (notify): Handle new notification types as 'D' for now, to allow testing
    this change.

* subversion/tests/cmdline/switch_tests.py
  (tree_conflicts_on_switch_1_1,
   tree_conflicts_on_switch_1_2): Update expected switch output.
* subversion/tests/cmdline/update_tests.py
  (tree_conflicts_on_switch_1_1,
   tree_conflicts_on_update_1_2): Update expected update output.