Remove the #ifdef SVN_DEBUG from the test functions in debug.c, to allow the linker to find these functions in release builds.
These functions are only for development use, so no references to these
functions should ever exist in our code.

The reason for this commit is that our Windows build environment declares
that the not static functions in this file should be exported as DLL
functions. But for this to work they must be available in release versions.

* subversion/libsvn_subr/debug.c
  (global): Remove #ifdef SVN_DEBUG around this file. Add some more details
    to the top level comment about why this code should not be used in
    release builds.