Add a working copy upgrade migration routine for moving properties from disk to the wc_db.
This commit does *not* change the current version number of the working copy,
so this code will never be run.  In fact, even with the current version number
bumped, I still can't get the code to run, for reasons I'll pontificate on
at a later date.  I'm committing this now to get some review.

Once we feel confident that the upgrade works as intended, I plan to commit
the rest of the patch: bumping the working copy version and using in-database
properties exclusively.

* subversion/libsvn_wc/upgrade.c
  (PROPS_SUBDIR, PROP_BASE_SUBDIR, TEMP_DIR): Some useful defines.
  (migrate_props): New.
  (bump_database_to_16): Renamed to...
  (bump_database_to_17): ...this.
  (bump_to_16): Renamed to...
  (bump_to_17): ...this.
  (svn_wc__upgrade_sdb): Call the new properties migration code when moving
    from format 15 -> 16.