Print an appropiate git diff header if a path is moved.
The diff code does not yet detect moves. We don't have a way to
distinguish between a copied and moved path with the diff callbacks.
Further, since the diff is created on the fly, we can't easily detect
if a deleted path is part of a move or just a plain delete.

But... At least the printing part is ready! :)

* subversion/libsvn_client/diff.c
  (diff_content_changed): Add call to print_git_diff_header_moved().