Fix the Windows ra_serf buildbot failures by not installing existing pristine files again.
This resolves the symptoms of the buildbot failures, but not the real
cause of the test failures: somehow we keep the filehandles open when
using ra_serf, but not while using different ra layers.

But this same issue will also trigger when we start using multiple clients
on one working copy and by file access by background processes, so
reinstalling the same file should be avoided anyway.

* subversion/libsvn_wc/wc_db.c
  (svn_wc__db_pristine_install): If the file already exists, just delete
    the to be installed tempfile and return.