Following up on r964613, make our SQLite code capable of recovering from a failed rollback because some statements were still busy.
This patch ensures that a rollback is really a rollback. Without this patch
the transaction is still in progress after returning from
svn_sqlite__with_transaction which will show up as reading and updating
an in-memory database that will never be committed.

(r964613, without the added svn_sqlite__reset() is a nice way to
test this code.)

* subversion/include/svn_error_codes.h
  (SVN_ERR_SQLITE_BUSY): New error code. Wraps SQLITE_BUSY.
  (SVN_ERR_SQLITE_RESETTING_FOR_ROLLBACK): New error code. Specifically for
    diagnosing buggy SQLite usage in our codebase.

* subversion/libsvn_subr/sqlite.c
  (SQLITE_ERROR_CODE): Map SQLITE_BUSY to SVN_ERR_SQLITE_BUSY.
  (svn_sqlite__with_transaction): Handle SVN_ERR_SQLITE_BUSY to avoid
    continuing operations that will never be committed even though they
    think they don't use a transaction.