Make svn patch match a hunk the normal way around before trying to determine whether it has already been applied by matching it in reverse.
This fixes a bug where when applying a patch and then applying it again in
reverse, a hunk would be misdetected as "already applied", with the following
type of patch:

  Index: foo
  ===================================================================
  --- foo	(revision 2)
  +++ foo	(working copy)
  @@ -1 +1,2 @@
   alpha
  +aaa

This change has a side effect: Instead of causing a "hunk already applied"
message upon duplicate application, repeatedly applying the patch above will
now keep adding "aaa" lines to the file foo, as long as the first line of the
file reads "alpha". Conversely, applying the patch in reverse will keep
removing "aaa" lines following the "alpha" line lines until none are left.
However, this matches the behaviour of Larry Wall's patch implementation,
so we accept it for now. GNU patch detects the duplicate application,
just as Subversion did before this commit. But both Larry's and GNU's patch
implementations correctly apply the above patch in reverse, so the fix for
this has higher priority.

* subversion/tests/cmdline/patch_tests.py
  (patch_same_twice): Adjust test expectations to conform to the behaviour
   of Larry Wall's patch implementation.
  (patch_reverse_revert): New test, which applies a patch and then applies
   it again in reverse. This test covers the bug this commit is fixing.
  (test_list): Add new test.

* subversion/libsvn_client/patch.c
  (get_hunk_info): Try to match the hunk the normal way around before
   checking whether it's already been applied.