Attempt to fix issue #3665 ("Double update of versioned file externals corrupting file timestamps") by trying to better preserve the two-phase handling of working copies with externals in the file-external case, too.
My approach is two-fold.  First, teach the reporter driver to not
report file externals as switched items when they are crawled as part
of a crawl of their parent directory (as opposed to when they are
crawled explicitly).  This should cause the server to stop sending
updates for them.  Secondly, teach the update editor code to not claim
that file externals which were not the explicit target of the
update/switch were updated when, in fact, they weren't.

(I'm not particularly confident that this doesn't break something else
in a mysterious fashion.)

* subversion/libsvn_wc/update_editor.c
  (tweak_entries): Skip file external children.

* subversion/libsvn_wc/adm_crawler.c
  (report_revisions_and_depths): Skip file external children.