In an effort to actually cut a 1.6.x release, reverse merge r1033111, which was itself a merge of r962377 and r962378 from trunk.
This merge created a test failure of svnsync on bdb, and this failure was not
easily understood or fixed, so we're backing out the change in favor of
cutting a release, and pushing a proper fix for this to a future release.

stsp, pburba, hwright and danielsh were all in agreement on this.