Introduce a new helper function which compares two checksums and then returns the appropriate error.
One reason for this is to consolidate and consistify the various strings used
in the error message (to help translators).  Another is to remove (or at least
reduce) the occurance of the "format not a string literal, argument types not
checked" warnings generated by the old style of use.

Subsequent commits will update other uses and make the error message generation
a tad more flexible.

* subversion/libsvn_subr/checksum.c
  (svn_checksum_mismatch_err): New.

* subversion/include/svn_checksum.h
  (svn_checksum_mismatch_err): New.

* subversion/libsvn_client/export.c
  (close_file): Use the new API, rather than checking the checksum manually.