Introduce a new way of verifying error output in the python testsuite: by error code.
Since we now output the error code in our error messages, we can just look
for the code, rather than matching the (frequently changing) error text.
(This will also help if somebody wanted to run the tests in a locale other
than en.)

This commit also adds a script to generate the list of error codes in a
format useful to the tests.  For convenience, the output of that script is
also versioned.

* subversion/tests/cmdline/diff_tests.py
  (diff_wrong_extension_type): Look for an error by code, rather than text.

* subversion/tests/cmdline/svntest/err.py:
  New autogenerated script.

* subversion/tests/cmdline/svntest/verify.py
  (createExpectedOutput): Recognized an expected error, and create a
    corresponding ExpectedOutput object for it (a regex matching the code).

* tools/dev/gen-py-errors.py:
  New script to generate the error code list.