Ah, the foibles of youth.
Two years ago, I committed r875310 which fixed issue #2220:
"'svn propdel' returns success on deleting a non-existent prop".  It turns out
that my younger, less-experienced self was just making life hard for the
slightly more gray-haired and battle-hardened self I am today (no doubt caused
by code such as this).

So, in homage to the ever-continuing course of self-improvement (and in an
effort to over-use and even abuse the punctuation mark know as the hyphen), I
give you this revision, which simplifies, clarifies, and improve-ifies the
non-existent property deletion notification.

Just use the standard notification infrastructure to when we attempt to
delete a nonexistent property.

* subversion/tests/cmdline/prop_tests.py
  (invalid_propnames, perms_on_symlink, delete_nonexistent_property):
    Update test expectations.
 
* subversion/svn/propdel-cmd.c
  (notify_wrapper_baton, notify_wrapper): Remove.
  (svn_cl__propdel): Remove the notification wrappers and post-process check.

* subversion/svn/notify.c
  (notify): Handle the svn_wc_notify_property_deleted_nonexistent case.