Fix a bug when syncing file flags with properties.
In spite of the name, directories are perfectly valid targets of this work
queue action, and when doing so on *nix, interesting things can happen.
Because most directories don't have an explicit svn:executable property
set on them, invoking this action on a directory will *remove* the eXecute
bit, which causes all kinds of havoc for the unsuspecting user.

So, we just special case directories, by only removing the execute bits if
the node in question isn't a directory.  Arguably, this is a bit inefficient
as currently implemented, but at least it is Correct.

(And for good measure, don't even bother with the whole mess if we're on
Windows.)

* subversion/libsvn_wc/workqueue.c
  (sync_file_flags): Check to see if we're looking at a directory before
    attempting to remove the eXecute bit.