[This is patch 1/3. This one is self contained, but doesn't update   subversion/libsvn_wc/external.c, subversion/libsvn_wc/merge.c   and subversion/libsvn_wc/update_editor.c yet.]
Make the translate api use a passed set of properties and make retrieving the
values needed for keyword expansion optional for the simple normalization case.

This patch prepares this part of our libsvn_wc api for when file externals
are no longer normal versioned files.

The next patch updates the editors to use the new api and the last one will
update the merge code. You need all or none of these patches, but this allows
easier review.

* subversion/libsvn_wc/adm_ops.c
  (svn_wc__internal_remove_from_revision_control): Update caller.

* subversion/libsvn_wc/props.c
  (do_propset): Update caller.

* subversion/libsvn_wc/questions.c
  (compare_and_verify): Update caller.

* subversion/libsvn_wc/translate.c
  (svn_wc__internal_translated_stream): Update caller.
  (svn_wc__internal_translated_file): Update caller.
  (svn_wc__get_translate_info): Forward for_normalization boolean.
  (svn_wc__expand_keywords): If normalizing, don't read from db. Try to obtain
    url via read info.

* subversion/libsvn_wc/translate.h
  (svn_wc__get_translate_info): Update arguments and documentation.
  (svn_wc__expand_keywords): Update arguments and documentation.

* subversion/libsvn_wc/workqueue.c
  (install_committed_file,
   run_file_install,
   run_file_copy_translated): Update callers.