Following up on r1102242, improve svnversion performance by another 40-50% (tested on a hot filesystem cache. Result is at least better than before  in other scenarios.).
Instead of doing a broad query to retrieve all nodes and then call a modcheck
on a subset of these nodes, query only the file and their recorded information.
(And as we already checked that there are no working layers we can even limit
 it to op_depth 0, to remove another subquery per file).

With that information we can do a filestat and immediately stop checking
for further changes if the recorded information matches.

This avoids two queries for every unmodified file in a working copy that
is checked before the first modification is found.

* subversion/libsvn_wc/wc-queries.sql
  (STMT_SELECT_CURRENT_NODES_RECURSIVE): Rename to ...
  (STMT_SELECT_BASE_FILES_RECURSIVE): ... this and specialize the result
    for its only user.

* subversion/libsvn_wc/wc_db.c
  (has_local_mods): Don't filter on kind. Perform a recorded information
    check directly instead of only after we do a svn_wc__db_read_info().