Fix issue 3859: we missed providing a notification in certain situations.
Specifically, the triggers were marking rows in the 'revert_list' but were
working against each other. This change removes the triggers from one set
of code paths and manually inserts the appropriate rows.

There is more cleanup on this code path, and the "recursive revert" code
paths can (and should?) be switched away from triggers. But that work is
for a future revision.

* subversion/tests/cmdline/revert_tests.py:
  (revert_empty_actual): remove the XFail decorator

* subversion/libsvn_wc/wc-queries.sql:
  (STMT_INSERT_REVERT_LIST): insert an item into revert_list, picking up
    conflict information from the ACTUAL_NODE, if any
  (STMT_INSERT_REVERT_LIST_PLAIN): just jam a simple row in there
  (STMT_INSERT_REVERT_LIST_ACTUAL_ONLY): insert a row based on any
    available information from ACTUAL_NODE

* subversion/libsvn_wc/wc_db.c:
  (op_revert_txn): drop the triggers on entry. we don't want to use them.
    we'll just avoid the triggers in the future. for each appropriate
    condition of what we find in the tables, insert an appropriate row
    into the revert_list.