As part of issue #3779, "actual-only nodes need regression tests", make 'svn add' detect tree conflict victims that do not exist on disk and prevent adding new nodes at that path with a meaningful error message.
This implies that if users need to add a new node to resolve the conflict
they need to mark the conflict as resolved first. I think this is safer
than allowing accidental additions to take place. Since the node is not
visible on disk the addition might be a mistake.

* subversion/libsvn_wc/adm_ops.c
  (check_can_add_node): Don't allow adding new items on top of nonexistent
   conflicted nodes.

* subversion/libsvn_client/add.c
  (add): As previous.

* subversion/tests/cmdline/tree_conflict_tests.py
  (actual_only_node_behaviour): Adjust test cases for 'add' and 'mkdir'.