Tweak the svn_repos_notify_* API to be less tailored for svnadmin.
Note, since the notifications were printed both during the dump and once
more at the end, the API will now generate the same svn_repos_notify_action_t
twice: once during the dump and once at the end.  This should be fine for API
users since we have (or are about to have) a "Done" notification.

* subversion/include/svn_repos.h
  (svn_repos_notify_warning_t): New enum and typedef.
  (svn_repos_notify_t):
    Rename 'warning' to 'warning_str' and add 'warning' members.

* subversion/libsvn_repos/deprecated.c
  (repos_notify_handler): Track rename.

* subversion/libsvn_repos/dump.c
  (dump_node, svn_repos_dump_fs3): Track rename.
    Don't format the error message in an svnadmin-specific way.

* subversion/svnadmin/main.c
  (repos_notify_handler): Format the error message here.