Extract the heuristic that detects binary data from svn_io_detect_mimetype2() into a separate utility, svn_io_is_binary_data().
The greater plan is to detect whether data in svn: properties is binary.
We need this information when creating reject files for properties,
and also to offer more options during interactive conflict resolution of
property conflicts (currently only mine-full and theirs-full are supported,
which is suboptimal for multi-line properties like svn:ignore or svn:externals).

* subversion/include/svn_io.h
  (svn_io_is_binary_data): Declare.

* subversion/libsvn_subr/io.c
  (svn_io_detect_mimetype2): Call svn_io_is_binary_data().
  (svn_io_is_binary_data): New, extracted from svn_io_detect_mimetype2().