Track rename of the --pre-1.7-compatible option.
This record-only-merges r1145519 from trunk, which removed the option there
prior to the 1.7.0 release.  (See also r1145515.)

The test harness has to allow SERVER_MINOR_VERSION=7, as that is used for the
server_has_atomic_revprop() helper.  It may or may not accept
SERVER_MINOR_VERSION=8 at this time, but if it doesn't then that should be
fixed (eventually) on trunk, rather than on this branch.

* subversion/tests/cmdline/svntest/main.py
  (create_repos): Track rename.
  (server_has_fsfs_revprop_packing): New, unused.