Remove libsvn_auth_gpg_agent and move gpg-agent support into libsvn_subr.
There is no need to put gpg-agent support into a library loaded at runtime.
The svn code talks to the GPG agent via a UNIX stream socket.
There are no massive external library dependencies involved that
we need to keep at a distance, as with GNOME or KDE.

This should also fix the problem where svn fails to enable gpg-agent
support on Mac OS X because of a filename mismatch (*.so vs. *.dylib).

* subversion/include/svn_auth.h
  (svn_auth_gpg_agent_version): Remove declaration. A version function
   is only needed for DSOs.
  (svn_auth_get_gpg_agent_simple_provider): Declare this function on Darwin.
   Remove mention of libsvn_auth_gpg_agent from docstring.

* subversion/libsvn_auth_gpg_agent,
  subversion/libsvn_auth_gpg_agent/version.c: Remove.

* subversion/libsvn_auth_gpg_agent/gpg_agent.c: Move to ...

* subversion/libsvn_subr/gpg_agent.c: ... here. Wrap all code into
   the SVN_HAVE_GPG_AGENT define.

* subversion/libsvn_subr/auth.c
  (svn_auth_get_platform_specific_provider): Directly call
   svn_auth_get_gpg_agent_simple_provider() if SVN_HAVE_GPG_AGENT is defined
   instead of loading gpg-agent support as a DSO.

* subversion/bindings/swig/core.i: Remove svn_auth_gpg_agent_version.

* configure.ac,
  build.conf,
  build/generator/gen_make.py,
  build/generator/gen_win.py: Remove references to libsvn_auth_gpg_agent.