Fix a bug getting moved-to information for nested moves, i.e.:   svn mv a/b bb        |      svn mv a aa   svn delete a        O R     svn mv aa/b bb   svn status a/b       |      svn status a/b
Previously, scan_deletion("a/b") would first grab the proper "a/b moved to bb"
information, but would then carry on using the moved-to of node 'a' (again),
resulting in wrong returned values.

This patch causes 'svn status' to show correct moved-to info on nodes like
'a/b' above, that are the root of a move while "hidden" inside a
moved-away/deleted subtree.

* subversion/libsvn_wc/wc_db.c
  (scan_deletion_txn): Get moved-to info at most once. Also, if possible,
    break the loop with the first moved-to info found.