On the fs-successor-ids branch, switch write_final_rev() to dual-pool paradigm.
This fixes a pool-lifetime bug specific to the fs-successor-ids branch.
Successor-IDs from recursive calls (i.e. from subdirectories of the root)
were put into an iteration pool which was cleared before the results
were returned. This problem might eventually be resolved in a different way
since successors may in the future be stored in a temporary file associated
with the transaction, rather than being computed on the fly and stored in
memory all at once. But that will be done later.

* subversion/libsvn_fs_fs/fs_fs.c
  (write_final_rev): Switch to dual-pool paradigm
   Allocate values returned in the RESULT_POOL, and everything else in
   the SCRATCH_POOL. This function also takes a REPS_POOL parameter, the
   use of which is unchanged.
  (commit_body, commit_obliteration_body): Update callers.