On the fs-successor-ids branch, fix inconsistencies in how revision ranges are assigned to files in db/successors/*.
Successor data for revision zero is now created at repository-creation time.
No successors of anything are created in revision zero, so its successor
data is always empty.

This lets us assign revision-ranges to files as follows:

  0..999 => 0
  1000..1999 => 1
  etc.

instead of the previous:

  1..1000 => 0
  1001..2000 => 1
  etc.

which was documented, but did not match the actual behaviour of the code
in all places.

* BRANCH-README: Update documentation explaining how revision ranges are
   mapped onto successor files.

* subversion/libsvn_fs_fs/fs_fs.c
  (make_successor_ids_dirs): Create successor data for revision zero.
  (FSFS_SUCCESSORS_MAX_REVS_PER_FILE, update_successor_ids_file,
   update_successor_revisions_file): Don't special-case revision zero.
   This isn't necessary anymore, and was inconsistent with how the
   various path_successor_* helper functions behave.