Introduce a --enable-sqlite-compatibility-version=X.Y.Z option for ./configure to allow people building Subversion to specify how old the system being deployed on might be.  Setting the compatibility version to an older version turns on code in subversion that works around infelicities in older versions and relaxes the version check ("SQLite compiled for 3.7.4, but running with 3.7.3") at initialization time.
If the compat version is set to a version before the minimum supported
version (3.6.18), the build will fail early (at 'make') so the person
building can adjust her expectations.

The default for the compat version is the currently installed version,
so there should be no change in behavior for users not passing this
option to the configure script.


* subversion/include/private/svn_dep_compat.h
  (SVN_SQLITE_MIN_VERSION_NUMBER, SVN_SQLITE_MIN_VERSION): Set to
    SQLITE_VERSION_NUMBER, SQLITE_VERSION (respectively) if undefined.
  (SQLITE_VERSION_AT_LEAST): Check SVN_SQLITE_MIN_VERSION_NUMBER instead
    of SQLITE_VERSION_NUMBER.

* subversion/libsvn_subr/sqlite.c
  (init_sqlite):
    Check sqlite version against SVN_SQLITE_MIN_VERSION_NUMBER instead
    of SQLITE_VERSION_NUMBER.

* configure.ac:
    (--enable-sqlite-compatibility-version): New switch; sets
    SVN_SQLITE_MIN_VERSION_NUMBER and SVN_SQLITE_MIN_VERSION.

* notes/knobs
  (SVN_SQLITE_MIN_VERSION_NUMBER, SVN_SQLITE_MIN_VERSION): Document.


Patch by: Jonathan Nieder <jrnieder@gmail.com>
(I wrote the notes/knobs part (having forgotten to advise Jonathan about
that) and made minor tweaks to the remainder)