On the moves-scan-log branch, allow the heuristic to consider multiple move destinations for one deleted node. The user will have to resolve ambiguity.
For example, in situations such as:
  r4:
    D  foo
    A  bar (copied from foo:r3)
    A  baz (copied from foo:r3)
    A  bay (copied from foo:r3)
the user can tell svn to interpret one of the possible copy/delete pairs
as a move and treat the remaining copies as normal copies.

Note that the conflict callback implementation already supports this scenario.
But until now the library never returned more than one move suggestion.

* subversion/libsvn_wc/update_editor.c
  (find_applicable_moves): Don't stop processing moves from a single revision
   when one move was found. Continue processing in case there are more then
   one possible moves.

* subversion/libsvn_client/update.c
  (scan_moves_log_receiver): Remember all copies with the same source path
    as possible move destinations, instead treating them all as normal copies.