On the fs-py branch: Bring up-to-date with trunk.
("Why?" you may ask.  The simple answer is "because I can," the the longer
answer is that this is still a backburnered pet project of mine, and when I
decide to hack a bit more on it, I'll want it up-to-date.  Because it needs to
track changes in the FSFS backend, and a few of those have been made in the
past few weeks, it's wise to do the catchup merge now, while those changes are
still fresh in our collective heads.)