Ev2 shims: Fix the rest of the lock test failures.
This is a hack upon a hack.  The way we say "unlock" to a working copy is
by sending a prop delete on a non-existant property.  But since that property
doesn't exist in the first place, we weren't registering the delete correctly.
To "fix" the problem, we inject a false lock token when running update, and
then see if it's been deleted.

This is quite inefficient, and hopefully won't live too long (this is supposed
to be compatibility code, after all), but given Ev2's design, I'm not really
sure what the long term solution is.

Current number of failing Ev2 tests: 86

* subversion/libsvn_wc/update_editor.c
  (fetch_props_func): New.
  (make_editor): Use the file-private function to fetch props.