The conditional around the declaration of svn_auth_set_gnome_keyring_unlock_prompt_func (introduced in r1241585) was not functioning properly, it prevented the bindings from being generated on all platforms instead of just those without gnome-keyring support.
This is because the swig preprocessor did not have the needed #defines
available at the point where the function was declared. The reason the
conditional does work for the implementation, is that the implementation
is ignored by swig and included only in the core.c file, and that some
other include (presumably) makes svn_auth.h available.

To fix this, we move the declaration of the function to below the
include of svn_auth.h. While we here, also merge the declaration and
implementation using the swig %inline keyword.

* core.i
  (svn_auth_set_gnome_keyring_unlock_prompt_func): Merge declaration with
    implementation and move both below the include of svn_auth.h.

Patch by: Matthijs Kooijman <matthijs@stdin.nl>