Revert r1240999. This patch causes an unexpected behavior change on Windows, where before this patch the %PATH% variable was still available for hooks.
This unbreaks the svnlook tests on the Windows-RA buildbot.

After this patch Apr cleans the entire environment and svnlook.exe is unable
to find shared libraries like libapr-1.dll.

I think apr should be fixed to behave the same way on Windows as on unix
and we -as subversion project- should check if removing %PATH% on Windows
is a breaking change.

* subversion/libsvn_repos/hooks.c
  (env_from_env_hash): Revert to returning NULL for a NULL hash.