Tweak the recently introduced svn_stringbuf_create_empty() function.
Unlike with svn_string_create_empty() which produces an immutable string,
the resulting string buffer is often going to have data appended to it.  The
initial implementation created a non-standard buffer that claimed that no
space was allocated, not even for the null terminator, which was strictly
incorrect and meant it would always need reallocation before appending any
data.  By calling the standard create function, we ensure the class
invariants are maintained without any tricks, and also ensure there is
enough space for a little expansion before reallocation is required, without
sacrificing any significant space or speed.

* subversion/include/svn_string.h
  (svn_stringbuf_create_empty): Don't claim that the allocated size will be
    zero.

* subversion/libsvn_subr/svn_string.c
  (empty_buffer): Make this 'const' now that we can because it's only used
    for svn_string_t and not svn_stringbuf_t. Add a doc string.
  (create_stringbuf): Add a doc string and assertions.
  (svn_stringbuf_create_empty): Call svn_stringbuf_create_ensure() instead
    of constructing a non-standard, "very empty" stringbuf.