SvnAdmin should always have revprop caching enabled such that the infrastructure is being set up to notify *other* processes of changes done by e.g. setrevprop.
Even if the revprop update infrastructure fails for some reason,
the impact on svnadmin is low because it imposes a small overhead
for some operations but the process will always see its own changes.
Foreign changes may or may not become visible to the tool while
it is running -- just as today.

* subversion/svnadmin/main.c
  (open_repos): enable revprop caching