On the ev2-export branch: A first hack at commit support.
This fails miserably, because we don't yet install the pristine file during
the commit, and the post commit update process expects us to have done so.  I
hope to continue working on this problem.

There are also a few crashes in the delta path driver, which I think are
unrelated, and a result of recent changes on trunk.

* subversion/libsvn_client/commit_util.c
  (do_item_commit, svn_client__do_commit): Rewrite for Ev2.

* subversion/libsvn_client/commit.c
  (get_ra_editor): Return an Ev2 editor.
  (svn_client_commit6): Use an Ev2 editor to drive the commit.

* subversion/libsvn_client/client.h
  (svn_client__do_commit): Update params.

* subversion/libsvn_client/copy.c
  (wc_to_repos_copy): Use an Ev2 editor to drive the commit.