Make the status walker walk nodes in a deterministic order. This ensures sorted output of 'svn status'.
For the record, it would be nicer to sort in the 'svn' client rather than
within the library. But that is not possible without changing the
client<->library interface, which currently invokes a notification callback
per-node, rather than passing a list of all nodes.
See for related discussion: http://svn.haxx.se/dev/archive-2012-05/0245.shtml

* subversion/libsvn_wc/status.c
  (get_dir_status): Put elements of the child dirents hash table into a sorted
   list and walk the list instead of iterating over hash table elements.