Fix a problem with trying to fetch the Location: header when checking out a node. The code was examining HANDLER->DONE before it would extract the Location: header, but that was simply wrong. The Location: header is always set before an ra_serf response handler is invoked, so there is no need to check DONE. Further, when the DONE flag gets set, the response handler might never get called(!).
However, the handle_checkout() response handler is really just a cover
for expect_empty_body() with an "override" to capture the Location:
header. But the core ra_serf response processing *already* captures
the Location: header whenever it is present.

So... this is run synchronously, we can just examine the handler's
LOCATION field after the request completes. And we can use
expect_empty_body() directly.

* subversion/libsvn_ra_serf/commit.c:
  (checkout_context_t): not needed. removed.
  (create_checkout_body): the baton is now the activity_url.
  (handle_checkout): not needed. removed.
  (checkout_node): no longer set up a checkout_context_t, and just
    pass the activity_url as the BODY_DELEGATE_BATON. go ahead and use
    svn_ra_serf__expect_empty_body for the response handler. examine
    HANDLER.LOCATION after the request completes.