Remove the new in 1.8, but now unused tree conflict reason 'moved-away-and-edited'.
The moved-away status is just a cleaner way of reporting that the
node is deleted and at the same time the root moved. (Which could
also apply to a replaced node). The combination with edited applies
only to a slight subset of the cases and should really be reported
on the to-location of the move.

* subversion/bindings/javahl/src/org/apache/subversion/javahl/ConflictDescriptor.java
  (Reason): Sync with svn_wc.h.

* subversion/include/svn_wc.h
  (svn_wc_conflict_reason_t): Remove reason.

* subversion/libsvn_wc/conflicts.c
  (local_change_map): Remove mapping of reason.

* subversion/libsvn_wc/tree_conflicts.c
  (svn_wc__conflict_reason_map): Remove mapping of reason.

* subversion/libsvn_wc/update_editor.c
  (check_tree_conflict,
   delete_entry,
   open_directory,
   open_file): Remove check for unused tree conflict reason.

* subversion/svn/tree-conflicts.c
  (map_conflict_reason_human,
   map_conflict_reason_xml): Remove mapping of reason.